OpenXR.CStruct.Extends

Documentation

data BaseInStructure

XrBaseInStructure - Convenience type for iterating (read only)

Member Descriptions

Description

BaseInStructure can be used to facilitate iterating through a read-only structure pointer chain.

See Also

BaseInStructure, BaseOutStructure, StructureType

Constructors

BaseInStructure

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

next :: Ptr BaseInStructure

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR.

data BaseOutStructure

XrBaseOutStructure - Convenience type for iterating (mutable)

Member Descriptions

Description

BaseOutStructure can be used to facilitate iterating through a structure pointer chain that returns data back to the application.

See Also

BaseInStructure, BaseOutStructure, StructureType

Constructors

BaseOutStructure

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

next :: Ptr BaseOutStructure

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR.

type family Extends (a :: [Type] -> Type) (b :: Type) :: Constraint where ...

Equations

Extends CompositionLayerBaseHeader CompositionLayerColorScaleBiasKHR = ()

Extends CompositionLayerProjectionView CompositionLayerDepthInfoKHR = ()

Extends FrameEndInfo SecondaryViewConfigurationFrameEndInfoMSFT = ()

Extends FrameState SecondaryViewConfigurationFrameStateMSFT = ()

Extends HandJointLocationsEXT HandJointVelocitiesEXT = ()

Extends HandTrackerCreateInfoEXT HandPoseTypeInfoMSFT = ()

Extends InstanceCreateInfo InstanceCreateInfoAndroidKHR = ()

Extends InstanceCreateInfo DebugUtilsMessengerCreateInfoEXT = ()

Extends InteractionProfileSuggestedBinding InteractionProfileAnalogThresholdVALVE = ()

Extends InteractionProfileSuggestedBinding BindingModificationsKHR = ()

Extends SessionBeginInfo SecondaryViewConfigurationSessionBeginInfoMSFT = ()

Extends SessionCreateInfo GraphicsBindingOpenGLWin32KHR = ()

Extends SessionCreateInfo GraphicsBindingOpenGLXlibKHR = ()

Extends SessionCreateInfo GraphicsBindingOpenGLXcbKHR = ()

Extends SessionCreateInfo GraphicsBindingOpenGLWaylandKHR = ()

Extends SessionCreateInfo GraphicsBindingD3D11KHR = ()

Extends SessionCreateInfo GraphicsBindingD3D12KHR = ()

Extends SessionCreateInfo GraphicsBindingOpenGLESAndroidKHR = ()

Extends SessionCreateInfo GraphicsBindingVulkanKHR = ()

Extends SessionCreateInfo SessionCreateInfoOverlayEXTX = ()

Extends SessionCreateInfo GraphicsBindingEGLMNDX = ()

Extends SessionCreateInfo HolographicWindowAttachmentMSFT = ()

Extends SpaceLocation SpaceVelocity = ()

Extends SpaceLocation EyeGazeSampleTimeEXT = ()

Extends SwapchainCreateInfo SecondaryViewConfigurationSwapchainCreateInfoMSFT = ()

Extends SystemProperties SystemEyeGazeInteractionPropertiesEXT = ()

Extends SystemProperties SystemHandTrackingPropertiesEXT = ()

Extends SystemProperties SystemHandTrackingMeshPropertiesMSFT = ()

Extends ViewConfigurationView ViewConfigurationDepthRangeEXT = ()

Extends ViewConfigurationView ViewConfigurationViewFovEPIC = ()

Extends a b = TypeError ((ShowType a :<>: Text " is not extended by ") :<>: ShowType b)

class PeekChain es where

Methods

peekChain :: Ptr (Chain es) -> IO (Chain es)

class PokeChain es where

Methods

withChain :: Chain es -> (Ptr (Chain es) -> IO a) -> IO a

withZeroChain :: (Ptr (Chain es) -> IO a) -> IO a

type family Chain (xs :: [a]) = (r :: a) | r -> xs where ...

Equations

Chain '[] = ()

Chain (x ': xs) = (x, Chain xs)

type family Extendss (p :: [Type] -> Type) (xs :: [Type]) :: Constraint where ...

Equations

Extendss p '[] = ()

Extendss p (x ': xs) = (Extends p x, Extendss p xs)

data SomeStruct (a :: [Type] -> Type) where

Constructors

SomeStruct :: forall a es. (Extendss a es, PokeChain es, Show (Chain es)) => a es -> SomeStruct a

extendSomeStruct :: (Extensible a, Extends a e, ToCStruct e, Show e) => e -> SomeStruct a -> SomeStruct a

Add an extension to the beginning of the struct chain

This can be used to optionally extend structs based on some condition (for example, an extension or layer being available)

withSomeStruct :: forall a b. SomeStruct a -> (forall es. (Extendss a es, PokeChain es, Show (Chain es)) => a es -> b) -> b

Consume a SomeStruct value

withSomeCStruct :: forall a b. (forall es. (Extendss a es, PokeChain es) => ToCStruct (a es)) => SomeStruct a -> (forall es. (Extendss a es, PokeChain es) => Ptr (a es) -> IO b) -> IO b

Write the C representation of some extended a and use the pointer, the pointer must not be returned from the continuation.

peekSomeCStruct :: forall a. (Extensible a, forall es. (Extendss a es, PeekChain es) => FromCStruct (a es)) => Ptr (SomeStruct a) -> IO (SomeStruct a)

Given a pointer to a struct with an unknown chain, peek the struct and its chain.

pokeSomeCStruct

Arguments

:: (forall es. (Extendss a es, PokeChain es) => ToCStruct (a es))

=> Ptr (SomeStruct a)
Pointer to some memory at least the size of the head of the struct chain.

-> SomeStruct a
The struct to poke

-> IO b
Computation to run while the poked tail is valid

-> IO b

Given some memory for the head of the chain, allocate and poke the tail and run an action.

forgetExtensions :: Ptr (a es) -> Ptr (SomeStruct a)

Forget which extensions a pointed-to struct has by casting the pointer

class Extensible (a :: [Type] -> Type) where

Methods

extensibleTypeName :: String

For error reporting an invalid extension

getNext :: a es -> Chain es

setNext :: a ds -> Chain es -> a es

extends :: forall e b proxy. Typeable e => proxy e -> (Extends a e => b) -> Maybe b

pattern (::&) :: Extensible a => a es' -> Chain es -> a es infix 6

A pattern synonym to separate the head of a struct chain from the tail, use in conjunction with :& to extract several members.

Head{..} ::& () <- returningNoTail a b c
-- Equivalent to
Head{..} <- returningNoTail @'[] a b c

Head{..} ::& Foo{..} :& Bar{..} :& () <- returningWithTail a b c

myFun (Head{..} :&& Foo{..} :& ())

pattern (:&) :: e -> Chain es -> Chain (e ': es) infixr 7

View the head and tail of a Chain, see ::&

Equivalent to (,)

data SomeChild (a :: Type) where

Constructors

SomeChild :: forall a b. (Inherits a b, Typeable b, ToCStruct b, Show b) => b -> SomeChild a

withSomeChild :: SomeChild a -> (Ptr (SomeChild a) -> IO b) -> IO b

lowerChildPointer :: Inherits a b => Ptr b -> Ptr (SomeChild a)

type family Inherits (a :: Type) (b :: Type) :: Constraint where ...

Equations

Inherits SwapchainImageBaseHeader SwapchainImageD3D12KHR = ()

Inherits SwapchainImageBaseHeader SwapchainImageD3D11KHR = ()

Inherits SwapchainImageBaseHeader SwapchainImageVulkanKHR = ()

Inherits SwapchainImageBaseHeader SwapchainImageOpenGLESKHR = ()

Inherits SwapchainImageBaseHeader SwapchainImageOpenGLKHR = ()

Inherits (CompositionLayerBaseHeader '[]) CompositionLayerEquirect2KHR = ()

Inherits (CompositionLayerBaseHeader '[]) CompositionLayerEquirectKHR = ()

Inherits (CompositionLayerBaseHeader '[]) CompositionLayerCubeKHR = ()

Inherits (CompositionLayerBaseHeader '[]) CompositionLayerCylinderKHR = ()

Inherits (CompositionLayerBaseHeader '[]) CompositionLayerQuad = ()

Inherits (CompositionLayerBaseHeader '[]) CompositionLayerProjection = ()

Inherits HapticBaseHeader HapticVibration = ()

Inherits EventDataBaseHeader EventDataDisplayRefreshRateChangedFB = ()

Inherits EventDataBaseHeader EventDataMainSessionVisibilityChangedEXTX = ()

Inherits EventDataBaseHeader EventDataInteractionProfileChanged = ()

Inherits EventDataBaseHeader EventDataVisibilityMaskChangedKHR = ()

Inherits EventDataBaseHeader EventDataPerfSettingsEXT = ()

Inherits EventDataBaseHeader EventDataReferenceSpaceChangePending = ()

Inherits EventDataBaseHeader EventDataSessionStateChanged = ()

Inherits EventDataBaseHeader EventDataInstanceLossPending = ()

Inherits EventDataBaseHeader EventDataEventsLost = ()

Inherits LoaderInitInfoBaseHeaderKHR LoaderInitInfoAndroidKHR = ()

Inherits parent child = TypeError ((ShowType parent :<>: Text " is not inherited by ") :<>: ShowType child)

class Inheritable (a :: Type) where

Methods

peekSomeCChild :: Ptr (SomeChild a) -> IO (SomeChild a)