OpenXR.Core10.APIConstants

Documentation

pattern MIN_HAPTIC_DURATION :: forall a. Integral a => a

XR_MIN_HAPTIC_DURATION - Indicates the shortest valid duration for the device

See Also

applyHapticFeedback

type MAX_EXTENSION_NAME_SIZE = 128

pattern MAX_EXTENSION_NAME_SIZE :: forall a. Integral a => a

type MAX_API_LAYER_NAME_SIZE = 256

pattern MAX_API_LAYER_NAME_SIZE :: forall a. Integral a => a

type MAX_API_LAYER_DESCRIPTION_SIZE = 256

pattern MAX_API_LAYER_DESCRIPTION_SIZE :: forall a. Integral a => a

type MAX_SYSTEM_NAME_SIZE = 256

pattern MAX_SYSTEM_NAME_SIZE :: forall a. Integral a => a

type MAX_APPLICATION_NAME_SIZE = 128

pattern MAX_APPLICATION_NAME_SIZE :: forall a. Integral a => a

type MAX_ENGINE_NAME_SIZE = 128

pattern MAX_ENGINE_NAME_SIZE :: forall a. Integral a => a

type MAX_RUNTIME_NAME_SIZE = 128

pattern MAX_RUNTIME_NAME_SIZE :: forall a. Integral a => a

type MAX_PATH_LENGTH = 256

pattern MAX_PATH_LENGTH :: forall a. Integral a => a

type MAX_STRUCTURE_NAME_SIZE = 64

pattern MAX_STRUCTURE_NAME_SIZE :: forall a. Integral a => a

type MAX_RESULT_STRING_SIZE = 64

pattern MAX_RESULT_STRING_SIZE :: forall a. Integral a => a

type MAX_GRAPHICS_APIS_SUPPORTED = 32

pattern MAX_GRAPHICS_APIS_SUPPORTED :: forall a. Integral a => a

type MAX_ACTION_SET_NAME_SIZE = 64

pattern MAX_ACTION_SET_NAME_SIZE :: forall a. Integral a => a

type MAX_ACTION_NAME_SIZE = 64

pattern MAX_ACTION_NAME_SIZE :: forall a. Integral a => a

type MAX_LOCALIZED_ACTION_SET_NAME_SIZE = 128

pattern MAX_LOCALIZED_ACTION_SET_NAME_SIZE :: forall a. Integral a => a

type MAX_LOCALIZED_ACTION_NAME_SIZE = 128

pattern MAX_LOCALIZED_ACTION_NAME_SIZE :: forall a. Integral a => a

type MIN_COMPOSITION_LAYERS_SUPPORTED = 16

pattern MIN_COMPOSITION_LAYERS_SUPPORTED :: forall a. Integral a => a

type NULL_PATH = 0

pattern NULL_PATH :: forall a. Integral a => a

XR_NULL_PATH - A NULL semantic path

Description

The only https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrPath value defined to be constant across all instances is the invalid path NULL_PATH. No well-formed path string is associated with NULL_PATH. Unless explicitly permitted, it should not be passed to API calls or used as a structure attribute when a valid https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrPath is required.

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrPath

type NULL_SYSTEM_ID = 0

pattern NULL_SYSTEM_ID :: forall a. Integral a => a

XR_NULL_SYSTEM_ID - NULL system identifier

Description

The only https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId value defined to be constant across all instances is the invalid system NULL_SYSTEM_ID. No supported system is associated with NULL_SYSTEM_ID. Unless explicitly permitted, it should not be passed to API calls or used as a structure attribute when a valid https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId is required.

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, getSystem

type NO_DURATION = 0

pattern NO_DURATION :: forall a. Integral a => a

XR_NO_DURATION - Constant for no duration/immediate timeout

Description

For the case of timeout durations, NO_DURATION may be used to indicate that the timeout is immediate.

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrDuration

type INFINITE_DURATION = 9223372036854775807

pattern INFINITE_DURATION :: Int64

XR_INFINITE_DURATION - Constant for infinite duration/never times out

Description

INFINITE_DURATION is a special value that may be used to indicate that the timeout never occurs. A timeout with a duration that refers to the past has the same effect as a timeout of NO_DURATION.

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrDuration

type FREQUENCY_UNSPECIFIED = 0

pattern FREQUENCY_UNSPECIFIED :: forall a. Integral a => a

XR_FREQUENCY_UNSPECIFIED - Runtime should determine optimal frequency for haptic pulse

See Also

applyHapticFeedback

type MAX_EVENT_DATA_SIZE = 4016

pattern MAX_EVENT_DATA_SIZE :: Int

XR_MAX_EVENT_DATA_SIZE - Maximum event data buffer size

See Also

EventDataBuffer

type HAND_JOINT_COUNT_EXT = 26

pattern HAND_JOINT_COUNT_EXT :: forall a. Integral a => a

XR_HAND_JOINT_COUNT_EXT - The number of hand joint enums defined in XrHandJointEXT

Description

HAND_JOINT_COUNT_EXT defines the number of hand joint enumerants defined in HandJointEXT

See Also

HandJointEXT

type NULL_CONTROLLER_MODEL_KEY_MSFT = 0

pattern NULL_CONTROLLER_MODEL_KEY_MSFT :: forall a. Integral a => a

XR_NULL_CONTROLLER_MODEL_KEY_MSFT - The value representing an invalid model key

Description

NULL_CONTROLLER_MODEL_KEY_MSFT defines an invalid model key value.

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrControllerModelKeyMSFT

pattern VK_NULL_HANDLE :: IsHandle a => a

class (Eq a, Zero a) => IsHandle a

A class for things which can be created with VK_NULL_HANDLE.

class HasObjectType a where

Methods

objectTypeAndHandle :: a -> (ObjectType, Word64)

newtype Bool32

XrBool32 - Boolean value

Description

Boolean values used by OpenXR are of type https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32 and are 32-bits wide as suggested by the name. The only valid values are the following:

Enumerant Descriptions

See Also

ActionStateBoolean, ActionStateFloat, ActionStatePose, ActionStateVector2f, EventDataMainSessionVisibilityChangedEXTX, EventDataReferenceSpaceChangePending, FrameState, HandJointLocationsEXT, HandMeshMSFT, SecondaryViewConfigurationStateMSFT, SystemEyeGazeInteractionPropertiesEXT, SystemHandTrackingMeshPropertiesMSFT, SystemHandTrackingPropertiesEXT, SystemTrackingProperties, ViewConfigurationProperties, xrSetInputDeviceActiveEXT, xrSetInputDeviceStateBoolEXT

Constructors

Bool32 Int32

Bundled Patterns

pattern FALSE :: Bool32
FALSE represents a false value.

pattern TRUE :: Bool32
TRUE represents a true value.