OpenXR.Core10.Device

Documentation

getSystem

Arguments

:: forall io. MonadIO io

=> Instance
instance is the handle of the instance from which to get the information.

instance must be a valid Instance handle

-> SystemGetInfo
getInfo is a pointer to an SystemGetInfo structure containing the application’s requests for a system.

getInfo must be a pointer to a valid SystemGetInfo structure

-> io SystemId

xrGetSystem - Gets a system identifier

Parameter Descriptions

Description

To get an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, an application specifies its desired form factor to getSystem and gets the runtime’s https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId associated with that configuration.

If the form factor is supported but temporarily unavailable, getSystem must return ERROR_FORM_FACTOR_UNAVAILABLE. A runtime may return SUCCESS on a subsequent call for a form factor it previously returned ERROR_FORM_FACTOR_UNAVAILABLE. For example, connecting or warming up hardware might cause an unavailable form factor to become available.

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_FORM_FACTOR_UNAVAILABLE

ERROR_FORM_FACTOR_UNSUPPORTED

ERROR_VALIDATION_FAILURE

See Also

NULL_SYSTEM_ID, Instance, SystemGetInfo, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId

getSystemProperties

Arguments

:: forall a io. (Extendss SystemProperties a, PokeChain a, PeekChain a, MonadIO io)

=> Instance
instance is the instance from which systemId was retrieved.

instance must be a valid Instance handle

-> SystemId
systemId is the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId whose properties will be queried.

-> io (SystemProperties a)

xrGetSystemProperties - Gets the properties of a particular system

Parameter Descriptions

Description

An application can call getSystemProperties to retrieve information about the system such as vendor ID, system name, and graphics and tracking properties.

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_OUT_OF_MEMORY

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, SystemProperties

createSession

Arguments

:: forall a io. (Extendss SessionCreateInfo a, PokeChain a, MonadIO io)

=> Instance
instance is the instance from which systemId was retrieved.

instance must be a valid Instance handle

-> SessionCreateInfo a
createInfo is a pointer to an SessionCreateInfo structure containing information about how to create the session.

createInfo must be a pointer to a valid SessionCreateInfo structure

-> io Session

xrCreateSession - Creates an XrSession

Parameter Descriptions

Description

Creates a session using the provided createInfo and returns a handle to that session. This session is created in the SESSION_STATE_IDLE state, and a corresponding EventDataSessionStateChanged event to the SESSION_STATE_IDLE state must be generated as the first such event for the new session.

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_OUT_OF_MEMORY

ERROR_LIMIT_REACHED

ERROR_INITIALIZATION_FAILED

ERROR_SYSTEM_INVALID

ERROR_GRAPHICS_DEVICE_INVALID

ERROR_VALIDATION_FAILURE

ERROR_GRAPHICS_REQUIREMENTS_CALL_MISSING

See Also

ExtensionProperties, Instance, Session, SessionCreateFlags, SessionCreateInfo, beginSession, destroySession, endSession

withSession :: forall a io r. (Extendss SessionCreateInfo a, PokeChain a, MonadIO io) => Instance -> SessionCreateInfo a -> (io Session -> (Session -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createSession and destroySession

To ensure that destroySession is always called: pass bracket (or the allocate function from your favourite resource management library) as the last argument. To just extract the pair pass (,) as the last argument.

destroySession

Arguments

:: forall io. MonadIO io

=> Session
session is the session to destroy.

-> io ()

xrDestroySession - Destroys an XrSession

Parameter Descriptions

Description

Session handles are destroyed using destroySession. When an Session is destroyed, all handles that are children of that Session are also destroyed.

The application is responsible for ensuring that it has no calls using session in progress when the session is destroyed.

destroySession can be called when the session is in any session state.

Valid Usage (Implicit)

session must be a valid Session handle

Thread Safety

Access to session, and any child handles, must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

See Also

Session, beginSession, createSession, endSession

enumerateEnvironmentBlendModes :: forall io. MonadIO io => Instance -> SystemId -> ViewConfigurationType -> io ("environmentBlendModes" ::: Vector EnvironmentBlendMode)

xrEnumerateEnvironmentBlendModes - Lists environment blend modes

Parameter Descriptions

instance is the instance from which systemId was retrieved.

systemId is the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId whose environment blend modes will be enumerated.

viewConfigurationType is the ViewConfigurationType to enumerate.

environmentBlendModeCapacityInput is the capacity of the environmentBlendModes array, or 0 to indicate a request to retrieve the required capacity.

environmentBlendModeCountOutput is a pointer to the count of environmentBlendModes written, or a pointer to the required capacity in the case that environmentBlendModeCapacityInput is 0.

environmentBlendModes is a pointer to an array of EnvironmentBlendMode values, but can be NULL if environmentBlendModeCapacityInput is 0.

See Buffer Size Parameters chapter for a detailed description of retrieving the required environmentBlendModes size.

Description

Enumerates the set of environment blend modes that this runtime supports for a given view configuration of the system. Environment blend modes should be in order from highest to lowest runtime preference.

Runtimes must always return identical buffer contents from this enumeration for the given systemId and viewConfigurationType for the lifetime of the instance.

Valid Usage (Implicit)

instance must be a valid Instance handle

viewConfigurationType must be a valid ViewConfigurationType value

environmentBlendModeCountOutput must be a pointer to a uint32_t value

If environmentBlendModeCapacityInput is not 0, environmentBlendModes must be a pointer to an array of environmentBlendModeCapacityInput EnvironmentBlendMode values

Return Codes

Success

SUCCESS

Failure

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_VIEW_CONFIGURATION_TYPE_UNSUPPORTED

ERROR_SIZE_INSUFFICIENT

See Also

EnvironmentBlendMode, Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, ViewConfigurationType

newtype SystemId

XrSystemId - Identifier for a system

Description

An https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId is an opaque atom used by the runtime to identify a system. The value NULL_SYSTEM_ID is considered an invalid system.

See Also

NULL_SYSTEM_ID, SessionCreateInfo, SystemProperties, VulkanDeviceCreateInfoKHR, VulkanGraphicsDeviceGetInfoKHR, VulkanInstanceCreateInfoKHR, enumerateEnvironmentBlendModes, enumerateViewConfigurationViews, enumerateViewConfigurations, getD3D11GraphicsRequirementsKHR, getD3D12GraphicsRequirementsKHR, getOpenGLESGraphicsRequirementsKHR, getOpenGLGraphicsRequirementsKHR, getSystem, getSystemProperties, getViewConfigurationProperties, getVulkanDeviceExtensionsKHR, getVulkanGraphicsDeviceKHR, getVulkanGraphicsRequirements2KHR, getVulkanGraphicsRequirementsKHR, getVulkanInstanceExtensionsKHR

Constructors

SystemId Word64

data SystemGetInfo

XrSystemGetInfo - Specifies desired attributes of the system

Member Descriptions

Description

The SystemGetInfo structure specifies attributes about a system as desired by an application.

Valid Usage (Implicit)

See Also

FormFactor, StructureType, getSystem

Constructors

SystemGetInfo

Fields

formFactor :: FormFactor

formFactor is the FormFactor requested by the application.

formFactor must be a valid FormFactor value

data SystemProperties (es :: [Type])

XrSystemProperties - Properties of a particular system

Valid Usage (Implicit)

See Also

StructureType, SystemGraphicsProperties, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, SystemTrackingProperties, getSystem, getSystemProperties

Constructors

SystemProperties

Fields

next :: Chain es

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR.

next must be NULL or a valid pointer to the next structure in a structure chain. See also: SystemEyeGazeInteractionPropertiesEXT, SystemHandTrackingMeshPropertiesMSFT, SystemHandTrackingPropertiesEXT

systemId :: SystemId

systemId is the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId identifying the system.

vendorId :: Word32

vendorId is a unique identifier for the vendor of the system.

systemName :: ByteString

systemName is a string containing the name of the system.

graphicsProperties :: SystemGraphicsProperties

graphicsProperties is an SystemGraphicsProperties structure specifying the system graphics properties.

trackingProperties :: SystemTrackingProperties

trackingProperties is an SystemTrackingProperties structure specifying system tracking properties.

data SystemGraphicsProperties

XrSystemGraphicsProperties - Graphics-related properties of a particular system

Member Descriptions

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, SystemProperties, SystemTrackingProperties, getSystem, getSystemProperties

Constructors

SystemGraphicsProperties

Fields

maxSwapchainImageHeight :: Word32

maxSwapchainImageHeight is the maximum swapchain image pixel height supported by this system.

maxSwapchainImageWidth :: Word32

maxSwapchainImageWidth is the maximum swapchain image pixel width supported by this system.

maxLayerCount :: Word32

maxLayerCount is the maximum number of composition layers supported by this system. The runtime must support at least MIN_COMPOSITION_LAYERS_SUPPORTED layers.

data SystemTrackingProperties

XrSystemTrackingProperties - Tracking-related properties of a particular system

Member Descriptions

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32, SystemGraphicsProperties, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, SystemProperties, getSystem, getSystemProperties

Constructors

SystemTrackingProperties

Fields

orientationTracking :: Bool

orientationTracking is set to TRUE to indicate the system supports orientational tracking of the view pose(s), FALSE otherwise.

positionTracking :: Bool

positionTracking is set to TRUE to indicate the system supports positional tracking of the view pose(s), FALSE otherwise.

data SessionCreateInfo (es :: [Type])

XrSessionCreateInfo - Creates a session

Valid Usage (Implicit)

See Also

SessionCreateFlags, StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, createSession

Constructors

SessionCreateInfo

Fields

next :: Chain es

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR. Note that in most cases one graphics API extension specific struct needs to be in this next chain.

next, unless otherwise specified via an extension, must contain exactly one graphics API binding structure (a structure whose name begins with "XrGraphicsBinding") or ERROR_GRAPHICS_DEVICE_INVALID must be returned.

next must be NULL or a valid pointer to the next structure in a structure chain. See also: GraphicsBindingD3D11KHR, GraphicsBindingD3D12KHR, GraphicsBindingEGLMNDX, GraphicsBindingOpenGLESAndroidKHR, GraphicsBindingOpenGLWaylandKHR, GraphicsBindingOpenGLWin32KHR, GraphicsBindingOpenGLXcbKHR, GraphicsBindingOpenGLXlibKHR, GraphicsBindingVulkanKHR, HolographicWindowAttachmentMSFT, SessionCreateInfoOverlayEXTX

createFlags :: SessionCreateFlags

createFlags identifies SessionCreateFlags that apply to the creation.

createFlags must be 0

systemId :: SystemId

systemId is the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId representing the system of devices to be used by this session.

systemId must be a valid https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId or ERROR_SYSTEM_INVALID must be returned.