OpenXR.Core10.Enums.EnvironmentBlendMode

Documentation

newtype EnvironmentBlendMode

XrEnvironmentBlendMode - Environment blend modes

Enumerant Descriptions

See Also

FrameEndInfo, SecondaryViewConfigurationLayerInfoMSFT, enumerateEnvironmentBlendModes

Constructors

EnvironmentBlendMode Int32

Bundled Patterns

pattern ENVIRONMENT_BLEND_MODE_OPAQUE :: EnvironmentBlendMode
ENVIRONMENT_BLEND_MODE_OPAQUE. The composition layers will be displayed with no view of the physical world behind them. The composited image will be interpreted as an RGB image, ignoring the composited alpha channel. This is the typical mode for VR experiences, although this mode can also be supported on devices that support video passthrough.

pattern ENVIRONMENT_BLEND_MODE_ADDITIVE :: EnvironmentBlendMode
ENVIRONMENT_BLEND_MODE_ADDITIVE. The composition layers will be additively blended with the real world behind the display. The composited image will be interpreted as an RGB image, ignoring the composited alpha channel during the additive blending. This will cause black composited pixels to appear transparent. This is the typical mode for an AR experience on a see-through headset with an additive display, although this mode can also be supported on devices that support video passthrough.

pattern ENVIRONMENT_BLEND_MODE_ALPHA_BLEND :: EnvironmentBlendMode
ENVIRONMENT_BLEND_MODE_ALPHA_BLEND. The composition layers will be alpha-blended with the real world behind the display. The composited image will be interpreted as an RGBA image, with the composited alpha channel determining each pixel’s level of blending with the real world behind the display. This is the typical mode for an AR experience on a phone or headset that supports video passthrough.