OpenXR.Core10.Enums.FormFactor

Documentation

newtype FormFactor

XrFormFactor - Supported form factors

Description

The predefined form factors which may be supported by OpenXR runtimes are:

Enumerant Descriptions

See Also

NULL_SYSTEM_ID, SystemGetInfo, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, getSystem

Constructors

FormFactor Int32

Bundled Patterns

pattern FORM_FACTOR_HEAD_MOUNTED_DISPLAY :: FormFactor
FORM_FACTOR_HEAD_MOUNTED_DISPLAY. The tracked display is attached to the user’s head. The user cannot touch the display itself. A VR headset would be an example of this form factor.

pattern FORM_FACTOR_HANDHELD_DISPLAY :: FormFactor
FORM_FACTOR_HANDHELD_DISPLAY. The tracked display is held in the user’s hand, independent from the user’s head. The user may be able to touch the display, allowing for screen-space UI. A mobile phone running an AR experience using pass-through video would be an example of this form factor.