OpenXR.Core10.Enums.Result

Documentation

pattern SUCCEEDED :: Result

XR_SUCCEEDED - Success convenience macro

Description

This may be a qualified success such as FRAME_DISCARDED.

See Also

FAILED, UNQUALIFIED_SUCCESS

pattern UNQUALIFIED_SUCCESS :: Result

XR_UNQUALIFIED_SUCCESS - Unqualified success convenience macro

Description

used to compare an Result to 0 (SUCCESS) exclusively.

See Also

FAILED, SUCCEEDED

pattern FAILED :: Result

XR_FAILED - Failure convenience macro

Description

some way.

See Also

SUCCEEDED, UNQUALIFIED_SUCCESS

newtype Result

XrResult - Result codes

Description

All return codes in the API are reported via Result return values.

Some common suffixes shared across many of the return codes are defined below:

See Also

resultToString

Constructors

Result Int32

Bundled Patterns

pattern SUCCESS :: Result

pattern TIMEOUT_EXPIRED :: Result

pattern SESSION_LOSS_PENDING :: Result

pattern EVENT_UNAVAILABLE :: Result

pattern SPACE_BOUNDS_UNAVAILABLE :: Result

pattern SESSION_NOT_FOCUSED :: Result

pattern FRAME_DISCARDED :: Result

pattern ERROR_VALIDATION_FAILURE :: Result

pattern ERROR_RUNTIME_FAILURE :: Result

pattern ERROR_OUT_OF_MEMORY :: Result

pattern ERROR_API_VERSION_UNSUPPORTED :: Result

pattern ERROR_INITIALIZATION_FAILED :: Result

pattern ERROR_FUNCTION_UNSUPPORTED :: Result

pattern ERROR_FEATURE_UNSUPPORTED :: Result

pattern ERROR_EXTENSION_NOT_PRESENT :: Result

pattern ERROR_LIMIT_REACHED :: Result

pattern ERROR_SIZE_INSUFFICIENT :: Result

pattern ERROR_HANDLE_INVALID :: Result

pattern ERROR_INSTANCE_LOST :: Result

pattern ERROR_SESSION_RUNNING :: Result

pattern ERROR_SESSION_NOT_RUNNING :: Result

pattern ERROR_SESSION_LOST :: Result

pattern ERROR_SYSTEM_INVALID :: Result

pattern ERROR_PATH_INVALID :: Result

pattern ERROR_PATH_COUNT_EXCEEDED :: Result

pattern ERROR_PATH_FORMAT_INVALID :: Result

pattern ERROR_PATH_UNSUPPORTED :: Result

pattern ERROR_LAYER_INVALID :: Result

pattern ERROR_LAYER_LIMIT_EXCEEDED :: Result

pattern ERROR_SWAPCHAIN_RECT_INVALID :: Result

pattern ERROR_SWAPCHAIN_FORMAT_UNSUPPORTED :: Result

pattern ERROR_ACTION_TYPE_MISMATCH :: Result

pattern ERROR_SESSION_NOT_READY :: Result

pattern ERROR_SESSION_NOT_STOPPING :: Result

pattern ERROR_TIME_INVALID :: Result

pattern ERROR_REFERENCE_SPACE_UNSUPPORTED :: Result

pattern ERROR_FILE_ACCESS_ERROR :: Result

pattern ERROR_FILE_CONTENTS_INVALID :: Result

pattern ERROR_FORM_FACTOR_UNSUPPORTED :: Result

pattern ERROR_FORM_FACTOR_UNAVAILABLE :: Result

pattern ERROR_API_LAYER_NOT_PRESENT :: Result

pattern ERROR_CALL_ORDER_INVALID :: Result

pattern ERROR_GRAPHICS_DEVICE_INVALID :: Result

pattern ERROR_POSE_INVALID :: Result

pattern ERROR_INDEX_OUT_OF_RANGE :: Result

pattern ERROR_VIEW_CONFIGURATION_TYPE_UNSUPPORTED :: Result

pattern ERROR_ENVIRONMENT_BLEND_MODE_UNSUPPORTED :: Result

pattern ERROR_NAME_DUPLICATED :: Result

pattern ERROR_NAME_INVALID :: Result

pattern ERROR_ACTIONSET_NOT_ATTACHED :: Result

pattern ERROR_ACTIONSETS_ALREADY_ATTACHED :: Result

pattern ERROR_LOCALIZED_NAME_DUPLICATED :: Result

pattern ERROR_LOCALIZED_NAME_INVALID :: Result

pattern ERROR_GRAPHICS_REQUIREMENTS_CALL_MISSING :: Result

pattern ERROR_COLOR_SPACE_UNSUPPORTED_FB :: Result

pattern ERROR_DISPLAY_REFRESH_RATE_UNSUPPORTED_FB :: Result

pattern ERROR_CONTROLLER_MODEL_KEY_INVALID_MSFT :: Result

pattern ERROR_SECONDARY_VIEW_CONFIGURATION_TYPE_NOT_ENABLED_MSFT :: Result

pattern ERROR_CREATE_SPATIAL_ANCHOR_FAILED_MSFT :: Result

pattern ERROR_ANDROID_THREAD_SETTINGS_FAILURE_KHR :: Result

pattern ERROR_ANDROID_THREAD_SETTINGS_ID_INVALID_KHR :: Result