OpenXR.Core10.Enums.StructureType

Documentation

newtype StructureType

XrStructureType - Values for type members of structs

Description

Most structures containing type members have a value of type matching the type of the structure, as described more fully in Valid Usage for Structure Types.

Note that all extension enums begin at the extension enum base of 110 (base 10). Each extension is assigned a block of 1000 enums, starting at the enum base and arranged by the extension’s index.

See Also

ActionCreateInfo, ActionSetCreateInfo, ActionSpaceCreateInfo, ActionStateBoolean, ActionStateFloat, ActionStateGetInfo, ActionStatePose, ActionStateVector2f, ActionsSyncInfo, ApiLayerProperties, BaseInStructure, BaseOutStructure, BindingModificationBaseHeaderKHR, BindingModificationsKHR, BoundSourcesForActionEnumerateInfo, CompositionLayerBaseHeader, CompositionLayerColorScaleBiasKHR, CompositionLayerCubeKHR, CompositionLayerCylinderKHR, CompositionLayerDepthInfoKHR, CompositionLayerEquirect2KHR, CompositionLayerEquirectKHR, CompositionLayerProjection, CompositionLayerProjectionView, CompositionLayerQuad, ControllerModelKeyStateMSFT, ControllerModelNodePropertiesMSFT, ControllerModelNodeStateMSFT, ControllerModelPropertiesMSFT, ControllerModelStateMSFT, DebugUtilsLabelEXT, DebugUtilsMessengerCallbackDataEXT, DebugUtilsMessengerCreateInfoEXT, DebugUtilsObjectNameInfoEXT, EventDataBaseHeader, EventDataBuffer, EventDataDisplayRefreshRateChangedFB, EventDataEventsLost, EventDataInstanceLossPending, EventDataInteractionProfileChanged, EventDataMainSessionVisibilityChangedEXTX, EventDataPerfSettingsEXT, EventDataReferenceSpaceChangePending, EventDataSessionStateChanged, EventDataVisibilityMaskChangedKHR, ExtensionProperties, EyeGazeSampleTimeEXT, FrameBeginInfo, FrameEndInfo, FrameState, FrameWaitInfo, GraphicsBindingD3D11KHR, GraphicsBindingD3D12KHR, GraphicsBindingEGLMNDX, GraphicsBindingOpenGLESAndroidKHR, GraphicsBindingOpenGLWaylandKHR, GraphicsBindingOpenGLWin32KHR, GraphicsBindingOpenGLXcbKHR, GraphicsBindingOpenGLXlibKHR, GraphicsBindingVulkanKHR, GraphicsRequirementsD3D11KHR, GraphicsRequirementsD3D12KHR, GraphicsRequirementsOpenGLESKHR, GraphicsRequirementsOpenGLKHR, GraphicsRequirementsVulkanKHR, HandJointLocationsEXT, HandJointVelocitiesEXT, HandJointsLocateInfoEXT, HandMeshMSFT, HandMeshSpaceCreateInfoMSFT, HandMeshUpdateInfoMSFT, HandPoseTypeInfoMSFT, HandTrackerCreateInfoEXT, HapticActionInfo, HapticBaseHeader, HapticVibration, HolographicWindowAttachmentMSFT, InputSourceLocalizedNameGetInfo, InstanceCreateInfo, InstanceCreateInfoAndroidKHR, InstanceProperties, InteractionProfileAnalogThresholdVALVE, InteractionProfileState, InteractionProfileSuggestedBinding, LoaderInitInfoAndroidKHR, LoaderInitInfoBaseHeaderKHR, ReferenceSpaceCreateInfo, SecondaryViewConfigurationFrameEndInfoMSFT, SecondaryViewConfigurationFrameStateMSFT, SecondaryViewConfigurationLayerInfoMSFT, SecondaryViewConfigurationSessionBeginInfoMSFT, SecondaryViewConfigurationStateMSFT, SecondaryViewConfigurationSwapchainCreateInfoMSFT, SessionActionSetsAttachInfo, SessionBeginInfo, SessionCreateInfo, SessionCreateInfoOverlayEXTX, SpaceLocation, SpaceVelocity, SpatialAnchorCreateInfoMSFT, SpatialAnchorSpaceCreateInfoMSFT, SpatialGraphNodeSpaceCreateInfoMSFT, SwapchainCreateInfo, SwapchainImageAcquireInfo, SwapchainImageBaseHeader, SwapchainImageD3D11KHR, SwapchainImageD3D12KHR, SwapchainImageOpenGLESKHR, SwapchainImageOpenGLKHR, SwapchainImageReleaseInfo, SwapchainImageVulkanKHR, SwapchainImageWaitInfo, SystemColorSpacePropertiesFB, SystemEyeGazeInteractionPropertiesEXT, SystemGetInfo, SystemHandTrackingMeshPropertiesMSFT, SystemHandTrackingPropertiesEXT, SystemProperties, View, ViewConfigurationDepthRangeEXT, ViewConfigurationProperties, ViewConfigurationView, ViewConfigurationViewFovEPIC, ViewLocateInfo, ViewState, VisibilityMaskKHR, VulkanDeviceCreateInfoKHR, VulkanGraphicsDeviceGetInfoKHR, VulkanInstanceCreateInfoKHR, VulkanSwapchainFormatListCreateInfoKHR, structureTypeToString

Constructors

StructureType Int32

Bundled Patterns

pattern TYPE_UNKNOWN :: StructureType

pattern TYPE_API_LAYER_PROPERTIES :: StructureType

pattern TYPE_EXTENSION_PROPERTIES :: StructureType

pattern TYPE_INSTANCE_CREATE_INFO :: StructureType

pattern TYPE_SYSTEM_GET_INFO :: StructureType

pattern TYPE_SYSTEM_PROPERTIES :: StructureType

pattern TYPE_VIEW_LOCATE_INFO :: StructureType

pattern TYPE_VIEW :: StructureType

pattern TYPE_SESSION_CREATE_INFO :: StructureType

pattern TYPE_SWAPCHAIN_CREATE_INFO :: StructureType

pattern TYPE_SESSION_BEGIN_INFO :: StructureType

pattern TYPE_VIEW_STATE :: StructureType

pattern TYPE_FRAME_END_INFO :: StructureType

pattern TYPE_HAPTIC_VIBRATION :: StructureType

pattern TYPE_EVENT_DATA_BUFFER :: StructureType

pattern TYPE_EVENT_DATA_INSTANCE_LOSS_PENDING :: StructureType

pattern TYPE_EVENT_DATA_SESSION_STATE_CHANGED :: StructureType

pattern TYPE_ACTION_STATE_BOOLEAN :: StructureType

pattern TYPE_ACTION_STATE_FLOAT :: StructureType

pattern TYPE_ACTION_STATE_VECTOR2F :: StructureType

pattern TYPE_ACTION_STATE_POSE :: StructureType

pattern TYPE_ACTION_SET_CREATE_INFO :: StructureType

pattern TYPE_ACTION_CREATE_INFO :: StructureType

pattern TYPE_INSTANCE_PROPERTIES :: StructureType

pattern TYPE_FRAME_WAIT_INFO :: StructureType

pattern TYPE_COMPOSITION_LAYER_PROJECTION :: StructureType

pattern TYPE_COMPOSITION_LAYER_QUAD :: StructureType

pattern TYPE_REFERENCE_SPACE_CREATE_INFO :: StructureType

pattern TYPE_ACTION_SPACE_CREATE_INFO :: StructureType

pattern TYPE_EVENT_DATA_REFERENCE_SPACE_CHANGE_PENDING :: StructureType

pattern TYPE_VIEW_CONFIGURATION_VIEW :: StructureType

pattern TYPE_SPACE_LOCATION :: StructureType

pattern TYPE_SPACE_VELOCITY :: StructureType

pattern TYPE_FRAME_STATE :: StructureType

pattern TYPE_VIEW_CONFIGURATION_PROPERTIES :: StructureType

pattern TYPE_FRAME_BEGIN_INFO :: StructureType

pattern TYPE_COMPOSITION_LAYER_PROJECTION_VIEW :: StructureType

pattern TYPE_EVENT_DATA_EVENTS_LOST :: StructureType

pattern TYPE_INTERACTION_PROFILE_SUGGESTED_BINDING :: StructureType

pattern TYPE_EVENT_DATA_INTERACTION_PROFILE_CHANGED :: StructureType

pattern TYPE_INTERACTION_PROFILE_STATE :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_ACQUIRE_INFO :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_WAIT_INFO :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_RELEASE_INFO :: StructureType

pattern TYPE_ACTION_STATE_GET_INFO :: StructureType

pattern TYPE_HAPTIC_ACTION_INFO :: StructureType

pattern TYPE_SESSION_ACTION_SETS_ATTACH_INFO :: StructureType

pattern TYPE_ACTIONS_SYNC_INFO :: StructureType

pattern TYPE_BOUND_SOURCES_FOR_ACTION_ENUMERATE_INFO :: StructureType

pattern TYPE_INPUT_SOURCE_LOCALIZED_NAME_GET_INFO :: StructureType

pattern TYPE_BINDING_MODIFICATIONS_KHR :: StructureType

pattern TYPE_SYSTEM_COLOR_SPACE_PROPERTIES_FB :: StructureType

pattern TYPE_EVENT_DATA_DISPLAY_REFRESH_RATE_CHANGED_FB :: StructureType

pattern TYPE_COMPOSITION_LAYER_EQUIRECT2_KHR :: StructureType

pattern TYPE_VULKAN_GRAPHICS_DEVICE_GET_INFO_KHR :: StructureType

pattern TYPE_VULKAN_DEVICE_CREATE_INFO_KHR :: StructureType

pattern TYPE_VULKAN_INSTANCE_CREATE_INFO_KHR :: StructureType

pattern TYPE_LOADER_INIT_INFO_ANDROID_KHR :: StructureType

pattern TYPE_INTERACTION_PROFILE_ANALOG_THRESHOLD_VALVE :: StructureType

pattern TYPE_HOLOGRAPHIC_WINDOW_ATTACHMENT_MSFT :: StructureType

pattern TYPE_VIEW_CONFIGURATION_VIEW_FOV_EPIC :: StructureType

pattern TYPE_CONTROLLER_MODEL_STATE_MSFT :: StructureType

pattern TYPE_CONTROLLER_MODEL_NODE_STATE_MSFT :: StructureType

pattern TYPE_CONTROLLER_MODEL_PROPERTIES_MSFT :: StructureType

pattern TYPE_CONTROLLER_MODEL_NODE_PROPERTIES_MSFT :: StructureType

pattern TYPE_CONTROLLER_MODEL_KEY_STATE_MSFT :: StructureType

pattern TYPE_SECONDARY_VIEW_CONFIGURATION_SWAPCHAIN_CREATE_INFO_MSFT :: StructureType

pattern TYPE_SECONDARY_VIEW_CONFIGURATION_LAYER_INFO_MSFT :: StructureType

pattern TYPE_SECONDARY_VIEW_CONFIGURATION_FRAME_END_INFO_MSFT :: StructureType

pattern TYPE_SECONDARY_VIEW_CONFIGURATION_FRAME_STATE_MSFT :: StructureType

pattern TYPE_SECONDARY_VIEW_CONFIGURATION_STATE_MSFT :: StructureType

pattern TYPE_SECONDARY_VIEW_CONFIGURATION_SESSION_BEGIN_INFO_MSFT :: StructureType

pattern TYPE_HAND_POSE_TYPE_INFO_MSFT :: StructureType

pattern TYPE_HAND_MESH_MSFT :: StructureType

pattern TYPE_HAND_MESH_UPDATE_INFO_MSFT :: StructureType

pattern TYPE_HAND_MESH_SPACE_CREATE_INFO_MSFT :: StructureType

pattern TYPE_SYSTEM_HAND_TRACKING_MESH_PROPERTIES_MSFT :: StructureType

pattern TYPE_HAND_JOINT_VELOCITIES_EXT :: StructureType

pattern TYPE_HAND_JOINT_LOCATIONS_EXT :: StructureType

pattern TYPE_HAND_JOINTS_LOCATE_INFO_EXT :: StructureType

pattern TYPE_HAND_TRACKER_CREATE_INFO_EXT :: StructureType

pattern TYPE_SYSTEM_HAND_TRACKING_PROPERTIES_EXT :: StructureType

pattern TYPE_SPATIAL_GRAPH_NODE_SPACE_CREATE_INFO_MSFT :: StructureType

pattern TYPE_GRAPHICS_BINDING_EGL_MNDX :: StructureType

pattern TYPE_VIEW_CONFIGURATION_DEPTH_RANGE_EXT :: StructureType

pattern TYPE_SPATIAL_ANCHOR_SPACE_CREATE_INFO_MSFT :: StructureType

pattern TYPE_SPATIAL_ANCHOR_CREATE_INFO_MSFT :: StructureType

pattern TYPE_COMPOSITION_LAYER_COLOR_SCALE_BIAS_KHR :: StructureType

pattern TYPE_EVENT_DATA_MAIN_SESSION_VISIBILITY_CHANGED_EXTX :: StructureType

pattern TYPE_SESSION_CREATE_INFO_OVERLAY_EXTX :: StructureType

pattern TYPE_EVENT_DATA_VISIBILITY_MASK_CHANGED_KHR :: StructureType

pattern TYPE_VISIBILITY_MASK_KHR :: StructureType

pattern TYPE_EYE_GAZE_SAMPLE_TIME_EXT :: StructureType

pattern TYPE_SYSTEM_EYE_GAZE_INTERACTION_PROPERTIES_EXT :: StructureType

pattern TYPE_GRAPHICS_REQUIREMENTS_D3D12_KHR :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_D3D12_KHR :: StructureType

pattern TYPE_GRAPHICS_BINDING_D3D12_KHR :: StructureType

pattern TYPE_GRAPHICS_REQUIREMENTS_D3D11_KHR :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_D3D11_KHR :: StructureType

pattern TYPE_GRAPHICS_BINDING_D3D11_KHR :: StructureType

pattern TYPE_GRAPHICS_REQUIREMENTS_VULKAN_KHR :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_VULKAN_KHR :: StructureType

pattern TYPE_GRAPHICS_BINDING_VULKAN_KHR :: StructureType

pattern TYPE_GRAPHICS_REQUIREMENTS_OPENGL_ES_KHR :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_OPENGL_ES_KHR :: StructureType

pattern TYPE_GRAPHICS_BINDING_OPENGL_ES_ANDROID_KHR :: StructureType

pattern TYPE_GRAPHICS_REQUIREMENTS_OPENGL_KHR :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_OPENGL_KHR :: StructureType

pattern TYPE_GRAPHICS_BINDING_OPENGL_WAYLAND_KHR :: StructureType

pattern TYPE_GRAPHICS_BINDING_OPENGL_XCB_KHR :: StructureType

pattern TYPE_GRAPHICS_BINDING_OPENGL_XLIB_KHR :: StructureType

pattern TYPE_GRAPHICS_BINDING_OPENGL_WIN32_KHR :: StructureType

pattern TYPE_DEBUG_UTILS_LABEL_EXT :: StructureType

pattern TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT :: StructureType

pattern TYPE_DEBUG_UTILS_MESSENGER_CALLBACK_DATA_EXT :: StructureType

pattern TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT :: StructureType

pattern TYPE_COMPOSITION_LAYER_EQUIRECT_KHR :: StructureType

pattern TYPE_COMPOSITION_LAYER_CYLINDER_KHR :: StructureType

pattern TYPE_EVENT_DATA_PERF_SETTINGS_EXT :: StructureType

pattern TYPE_VULKAN_SWAPCHAIN_FORMAT_LIST_CREATE_INFO_KHR :: StructureType

pattern TYPE_COMPOSITION_LAYER_DEPTH_INFO_KHR :: StructureType

pattern TYPE_INSTANCE_CREATE_INFO_ANDROID_KHR :: StructureType

pattern TYPE_COMPOSITION_LAYER_CUBE_KHR :: StructureType