OpenXR.Core10.Enums.ViewConfigurationType

Documentation

newtype ViewConfigurationType

XrViewConfigurationType - Supported view configuration type

Description

The application selects its primary view configuration type when calling beginSession, and that configuration remains constant for the lifetime of the session, until endSession is called.

The number of views and the semantic meaning of each view index within a given view configuration is well-defined, specified below for all core view configurations. The predefined primary view configuration types are:

Enumerant Descriptions

See Also

EventDataVisibilityMaskChangedKHR, SecondaryViewConfigurationLayerInfoMSFT, SecondaryViewConfigurationSessionBeginInfoMSFT, SecondaryViewConfigurationStateMSFT, SecondaryViewConfigurationSwapchainCreateInfoMSFT, SessionBeginInfo, ViewConfigurationProperties, ViewLocateInfo, enumerateEnvironmentBlendModes, enumerateViewConfigurationViews, enumerateViewConfigurations, getViewConfigurationProperties, getVisibilityMaskKHR

Constructors

ViewConfigurationType Int32

Bundled Patterns

pattern VIEW_CONFIGURATION_TYPE_PRIMARY_MONO :: ViewConfigurationType
VIEW_CONFIGURATION_TYPE_PRIMARY_MONO. One view representing the form factor’s one primary display. For example, an AR phone’s screen. This configuration requires one element in ViewConfigurationProperties and one projection in each CompositionLayerProjection layer.

pattern VIEW_CONFIGURATION_TYPE_PRIMARY_STEREO :: ViewConfigurationType
VIEW_CONFIGURATION_TYPE_PRIMARY_STEREO. Two views representing the form factor’s two primary displays, which map to a left-eye and right-eye view. This configuration requires two views in ViewConfigurationProperties and two views in each CompositionLayerProjection layer. View index 0 must represent the left eye and view index 1 must represent the right eye.

pattern VIEW_CONFIGURATION_TYPE_SECONDARY_MONO_FIRST_PERSON_OBSERVER_MSFT :: ViewConfigurationType

pattern VIEW_CONFIGURATION_TYPE_PRIMARY_QUAD_VARJO :: ViewConfigurationType