OpenXR.Core10.Haptics

Documentation

applyHapticFeedback

Arguments

:: forall a io. (ToCStruct a, MonadIO io)

=> Session
session is the Session to start outputting to.

session must be a valid Session handle

-> HapticActionInfo
hapticActionInfo is a pointer to HapticActionInfo to provide action and subaction paths information.

hapticActionInfo must be a pointer to a valid HapticActionInfo structure

-> ("hapticFeedback" ::: a)
hapticFeedback is a pointer to a haptic event structure which starts with an HapticBaseHeader.

hapticFeedback must be a pointer to a valid HapticBaseHeader-based structure. See also: HapticVibration

-> io Result

xrApplyHapticFeedback - Apply haptic feedback

Parameter Descriptions

Description

Triggers a haptic event through the specified action of type TYPE_HAPTIC_VIBRATION. The runtime should deliver this request to the appropriate device, but exactly which device, if any, this event is sent to is up to the runtime to decide. If an appropriate device is unavailable the runtime may ignore this request for haptic feedback.

If another haptic event from this session is currently happening on the device bound to this action, the runtime must interrupt that other event and replace it with the new one.

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_VALIDATION_FAILURE

ERROR_ACTIONSET_NOT_ATTACHED

ERROR_ACTION_TYPE_MISMATCH

ERROR_PATH_INVALID

ERROR_PATH_UNSUPPORTED

See Also

HapticActionInfo, HapticBaseHeader, HapticVibration, Session, stopHapticFeedback

stopHapticFeedback

Arguments

:: forall io. MonadIO io

=> Session
session is the Session to stop outputting to.

session must be a valid Session handle

-> HapticActionInfo
hapticActionInfo is a pointer to an HapticActionInfo to provide action and subaction path information.

hapticActionInfo must be a pointer to a valid HapticActionInfo structure

-> io Result

xrStopHapticFeedback - Stop haptic feedback

Parameter Descriptions

Description

If a haptic event from this Action is in progress, when this function is called the runtime must stop that event.

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_VALIDATION_FAILURE

ERROR_ACTIONSET_NOT_ATTACHED

ERROR_ACTION_TYPE_MISMATCH

ERROR_PATH_INVALID

ERROR_PATH_UNSUPPORTED

See Also

HapticActionInfo, Session, applyHapticFeedback

data HapticBaseHeader

XrHapticBaseHeader - Base header for haptic feedback

Valid Usage (Implicit)

See Also

HapticVibration, InteractionProfileAnalogThresholdVALVE, StructureType, applyHapticFeedback

Constructors

HapticBaseHeader

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

type must be TYPE_HAPTIC_VIBRATION

class ToCStruct a => IsHaptic a where

Methods

toHapticBaseHeader :: a -> HapticBaseHeader

data HapticActionInfo

XrHapticActionInfo - Information to output haptic feedback

Member Descriptions

Description

See ActionCreateInfo for a description of subaction paths, and the restrictions on their use.

Valid Usage (Implicit)

See Also

Action, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrPath, StructureType, applyHapticFeedback, stopHapticFeedback

Constructors

HapticActionInfo

Fields

action :: Ptr Action_T

action is the Action handle for the desired output haptic action.

action must be a valid Action handle

subactionPath :: Path

subactionPath is the subaction path https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrPath of the device to send the haptic event to, or NULL_PATH to specify all subaction paths. If the subaction path is specified, it is one of the subaction paths that were specified when the action was created. If the subaction path was not specified when the action was created, the runtime must return ERROR_PATH_UNSUPPORTED. If this parameter is specified, the runtime must trigger the haptic events only on the device from the subaction path.