OpenXR.Core10.OtherTypes

Documentation

data Vector4f

XrVector4f - Four-dimensional vector

Member Descriptions

Description

If used to represent physical distances, x, y, and z values must be in meters.

See Also

Posef, Quaternionf, Vector2f, Vector3f

Constructors

Vector4f

Fields

x :: Float

x is the x coordinate of the vector.

y :: Float

y is the y coordinate of the vector.

z :: Float

z is the z coordinate of the vector.

w :: Float

w is the w coordinate of the vector.

data Color4f

XrColor4f - Color Vector

Member Descriptions

Description

Unless otherwise specified, colors are encoded as linear (not with sRGB nor other gamma compression) values with individual components being in the range of 0.0 through 1.0, and without the RGB components being premultiplied by the alpha component.

See Also

CompositionLayerColorScaleBiasKHR

Constructors

Color4f

Fields

r :: Float

r is the red component of the color.

g :: Float

g is the green component of the color.

b :: Float

b is the blue component of the color.

a :: Float

a is the alpha component of the color.

data Fovf

XrFovf - Field of view

Member Descriptions

Description

Angles to the right of the center and upwards from the center are positive, and angles to the left of the center and down from the center are negative. The total horizontal field of view is angleRight minus angleLeft, and the total vertical field of view is angleUp minus angleDown. For a symmetric FoV, angleRight and angleUp will have positive values, angleLeft will be -angleRight, and angleDown will be -angleUp.

The angles must be specified in radians, and must be between -π/2 and π/2 exclusively.

When angleLeft > angleRight, the content of the view must be flipped horizontally. When angleDown > angleUp, the content of the view must be flipped vertically.

See Also

CompositionLayerProjectionView, View, ViewConfigurationViewFovEPIC

Constructors

Fovf

Fields

angleLeft :: Float

angleLeft is the angle of the left side of the field of view. For a symmetric field of view this value is negative.

angleRight :: Float

angleRight is the angle of the right side of the field of view.

angleUp :: Float

angleUp is the angle of the top part of the field of view.

angleDown :: Float

angleDown is the angle of the bottom part of the field of view. For a symmetric field of view this value is negative.

data SwapchainSubImage

XrSwapchainSubImage - Composition layer data

Valid Usage (Implicit)

See Also

CompositionLayerCylinderKHR, CompositionLayerDepthInfoKHR, CompositionLayerEquirect2KHR, CompositionLayerEquirectKHR, CompositionLayerProjectionView, CompositionLayerQuad, FrameEndInfo, Rect2Di, Swapchain

Constructors

SwapchainSubImage

Fields

swapchain :: Ptr Swapchain_T

swapchain is the Swapchain to be displayed.

swapchain must be a valid Swapchain handle

imageRect :: Rect2Di

imageRect is an Rect2Di representing the valid portion of the image to use, in pixels. It also implicitly defines the transform from normalized image coordinates into pixel coordinates. Note that the compositor may bleed in pixels from outside the bounds in some cases, for instance due to mipmapping.

imageArrayIndex :: Word32

imageArrayIndex is the image array index, with 0 meaning the first or only array element.

data CompositionLayerBaseHeader (es :: [Type])

XrCompositionLayerBaseHeader - Composition layer base header

Member Descriptions

Description

All composition layer structures begin with the elements described in the CompositionLayerBaseHeader. The CompositionLayerBaseHeader structure is not intended to be directly used, but forms a basis for defining current and future structures containing composition layer information. The FrameEndInfo structure contains an array of pointers to these polymorphic header structures. All composition layer type pointers must be type-castable as an CompositionLayerBaseHeader pointer.

Valid Usage (Implicit)

See Also

CompositionLayerFlags, FrameEndInfo, SecondaryViewConfigurationLayerInfoMSFT, Space, StructureType, SwapchainSubImage

Constructors

CompositionLayerBaseHeader

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

type must be one of the following XrStructureType values: TYPE_COMPOSITION_LAYER_CUBE_KHR, TYPE_COMPOSITION_LAYER_CYLINDER_KHR, TYPE_COMPOSITION_LAYER_EQUIRECT2_KHR, TYPE_COMPOSITION_LAYER_EQUIRECT_KHR, TYPE_COMPOSITION_LAYER_PROJECTION, TYPE_COMPOSITION_LAYER_QUAD

next :: Chain es

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR.

next must be NULL or a valid pointer to the next structure in a structure chain. See also: CompositionLayerColorScaleBiasKHR

layerFlags :: CompositionLayerFlags

layerFlags is a bitmask of XrCompositionLayerFlagBits describing flags to apply to the layer.

layerFlags must be 0 or a valid combination of XrCompositionLayerFlagBits values

space :: Ptr Space_T

space is the Space in which the layer will be kept stable over time.

space must be a valid Space handle

class ToCStruct a => IsCompositionLayer a where

Methods

toCompositionLayerBaseHeader :: a -> CompositionLayerBaseHeader '[]

data CompositionLayerProjectionView (es :: [Type])

XrCompositionLayerProjectionView - Projection layer element

Member Descriptions

Description

The count and order of view poses submitted with CompositionLayerProjection must be the same order as that returned by locateViews. The CompositionLayerProjectionView::pose and CompositionLayerProjectionView::fov should almost always derive from View::pose and View::fov as found in the locateViews::views array. However, applications may submit an CompositionLayerProjectionView which has a different view or FOV than that from locateViews. In this case, the runtime will map the view and FOV to the system display appropriately. In the case that two submitted views within a single layer overlap, they must be composited in view array order.

Valid Usage (Implicit)

See Also

CompositionLayerProjection, Fovf, Posef, StructureType, SwapchainSubImage

Constructors

CompositionLayerProjectionView

Fields

next :: Chain es

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR.

next must be NULL or a valid pointer to the next structure in a structure chain. See also: CompositionLayerDepthInfoKHR

pose :: Posef

pose is an Posef defining the location and orientation of this projection element in the space of the corresponding CompositionLayerProjectionView.

fov :: Fovf

fov is the Fovf for this projection element.

subImage :: SwapchainSubImage

subImage is the image layer SwapchainSubImage to use.

subImage must be a valid SwapchainSubImage structure

data CompositionLayerProjection

XrCompositionLayerProjection - Composition layer for projection

Member Descriptions

Description

Note

Because a runtime may reproject the layer over time, a projection layer should specify an Space in which to maximize stability of the layer content. For example, a projection layer containing world-locked content should use an Space which is also world-locked, such as the LOCAL or STAGE reference spaces. In the case that the projection layer should be head-locked, such as a heads up display, the VIEW reference space would provide the highest quality layer reprojection.

Valid Usage (Implicit)

type must be TYPE_COMPOSITION_LAYER_PROJECTION

next must be NULL or a valid pointer to the next structure in a structure chain

layerFlags must be 0 or a valid combination of XrCompositionLayerFlagBits values

space must be a valid Space handle

views must be a pointer to an array of viewCount valid CompositionLayerProjectionView structures

The viewCount parameter must be greater than 0

See Also

CompositionLayerBaseHeader, CompositionLayerFlags, CompositionLayerProjectionView, Space, StructureType, SwapchainSubImage

Constructors

CompositionLayerProjection

Fields

layerFlags :: CompositionLayerFlags

layerFlags is a bitmask of XrCompositionLayerFlagBits describing flags to apply to the layer.

space :: Ptr Space_T

space is the Space in which the pose of each CompositionLayerProjectionView is evaluated over time by the compositor.

views :: Vector (SomeStruct CompositionLayerProjectionView)

views is the array of type CompositionLayerProjectionView containing each projection layer view.

data CompositionLayerQuad

XrCompositionLayerQuad - Quad composition layer

Member Descriptions

Description

The CompositionLayerQuad layer is useful for user interface elements or 2D content rendered into the virtual world. The layer’s SwapchainSubImage::swapchain image is applied to a quad in the virtual world space. Only front face of the quad surface is visible; the back face is not visible and must not be drawn by the runtime. A quad layer has no thickness; it is a two-dimensional object positioned and oriented in 3D space. The position of a quad refers to the center of the quad within the given Space. The orientation of the quad refers to the orientation of the normal vector from the front face. The size of a quad refers to the quad’s size in the x-y plane of the given Space’s coordinate system. A quad with a position of {0,0,0}, rotation of {0,0,0,1} (no rotation), and a size of {1,1} refers to a 1 meter x 1 meter quad centered at {0,0,0} with its front face normal vector coinciding with the +z axis.

Valid Usage (Implicit)

See Also

CompositionLayerBaseHeader, CompositionLayerFlags, Extent2Df, EyeVisibility, Posef, Space, StructureType, SwapchainSubImage

Constructors

CompositionLayerQuad

Fields

layerFlags :: CompositionLayerFlags

layerFlags is a bitmask of XrCompositionLayerFlagBits describing flags to apply to the layer.

layerFlags must be 0 or a valid combination of XrCompositionLayerFlagBits values

space :: Ptr Space_T

space is the Space in which the pose of the quad layer is evaluated over time.

space must be a valid Space handle

eyeVisibility :: EyeVisibility

eyeVisibility is the EyeVisibility for this layer.

eyeVisibility must be a valid EyeVisibility value

subImage :: SwapchainSubImage

subImage is the image layer SwapchainSubImage to use.

subImage must be a valid SwapchainSubImage structure

pose :: Posef

pose is an Posef defining the position and orientation of the quad in the reference frame of the space.

size :: Extent2Df

size is the width and height of the quad in meters.

data HapticVibration

XrHapticVibration - Base header for haptic feedback

Member Descriptions

Description

The HapticVibration is used in calls to applyHapticFeedback that trigger vibration output actions.

The duration, and frequency parameters may be clamped to implementation-dependent ranges.

Valid Usage (Implicit)

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrDuration, HapticBaseHeader, StructureType, applyHapticFeedback

Constructors

HapticVibration

Fields

duration :: Duration

duration is the number of nanoseconds the vibration should last. If MIN_HAPTIC_DURATION is specified, the runtime must produce a short haptics pulse of minimal supported duration for the haptic device.

frequency :: Float

frequency is the frequency of the vibration in Hz. If FREQUENCY_UNSPECIFIED is specified, it is left to the runtime to decide the optimal frequency value to use.

amplitude :: Float

amplitude is the amplitude of the vibration between 0.0 and 1.0.

data EventDataBaseHeader

XrEventDataBaseHeader - Base header for an event

Parameter Descriptions

Description

The EventDataBaseHeader is a generic structure used to identify the common event data elements.

Upon receipt, the EventDataBaseHeader pointer should be type-cast to a pointer of the appropriate event data based on the type parameter.

Valid Usage (Implicit)

See Also

StructureType, pollEvent

Constructors

EventDataBaseHeader

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

type must be one of the following XrStructureType values: TYPE_EVENT_DATA_DISPLAY_REFRESH_RATE_CHANGED_FB, TYPE_EVENT_DATA_EVENTS_LOST, TYPE_EVENT_DATA_INSTANCE_LOSS_PENDING, TYPE_EVENT_DATA_INTERACTION_PROFILE_CHANGED, TYPE_EVENT_DATA_MAIN_SESSION_VISIBILITY_CHANGED_EXTX, TYPE_EVENT_DATA_PERF_SETTINGS_EXT, TYPE_EVENT_DATA_REFERENCE_SPACE_CHANGE_PENDING, TYPE_EVENT_DATA_SESSION_STATE_CHANGED, TYPE_EVENT_DATA_VISIBILITY_MASK_CHANGED_KHR

class ToCStruct a => IsEventData a where

Methods

toEventDataBaseHeader :: a -> EventDataBaseHeader

data EventDataEventsLost

XrEventDataEventsLost - Event indicating events were lost

Member Descriptions

Description

Receiving the EventDataEventsLost event structure indicates that the event queue overflowed and some events were removed at the position within the queue at which this event was found.

Valid Usage (Implicit)

See Also

EventDataBaseHeader, StructureType, pollEvent

Constructors

EventDataEventsLost

Fields

lostEventCount :: Word32

lostEventCount is the number of events which have overflowed since the last call to pollEvent.

data EventDataInstanceLossPending

XrEventDataInstanceLossPending - Event indicating instance loss will occur

Members

Receiving the EventDataInstanceLossPending event structure indicates that the application is about to lose the indicated Instance at the indicated lossTime in the future. The application should call destroyInstance and relinquish any instance-specific resources. This typically occurs to make way for a replacement of the underlying runtime, such as via a software update.

Description

After the application has destroyed all of its instances and their children and waited past the specified time, it may then re-try createInstance in a loop waiting for whatever maintenance the runtime is performing to complete. The runtime will return ERROR_INSTANCE_LOST from createInstance as long as it is unable to create the instance. Once the runtime has returned and is able to continue, it must resume returning SUCCESS from createInstance if valid data is passed in.

Valid Usage (Implicit)

See Also

EventDataBaseHeader, StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime, pollEvent

Constructors

EventDataInstanceLossPending

Fields

lossTime :: Time

lossTime is the absolute time at which the indicated instance will be considered lost and become unusable.

data EventDataSessionStateChanged

XrEventDataSessionStateChanged - Event indicating session state changed

Member Descriptions

Description

Receiving the EventDataSessionStateChanged event structure indicates that the application has changed lifecycle state.

Valid Usage (Implicit)

See Also

EventDataBaseHeader, Session, SessionState, StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime, pollEvent

Constructors

EventDataSessionStateChanged

Fields

session :: Ptr Session_T

session is the Session which has changed state.

session must be a valid Session handle

state :: SessionState

state is the current SessionState of the session.

state must be a valid SessionState value

time :: Time

time is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime which indicates the time of the state change.

data EventDataReferenceSpaceChangePending

XrEventDataReferenceSpaceChangePending - Notifies the application that a reference space is changing

Valid Usage (Implicit)

See Also

REFERENCE_SPACE_TYPE_STAGE, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32, Posef, ReferenceSpaceType, Session, StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime, createReferenceSpace

Constructors

EventDataReferenceSpaceChangePending

Fields

session :: Ptr Session_T

session is the Session for which the reference space is changing.

session must be a valid Session handle

referenceSpaceType :: ReferenceSpaceType

referenceSpaceType is the ReferenceSpaceType that is changing.

referenceSpaceType must be a valid ReferenceSpaceType value

changeTime :: Time

changeTime is the target https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime after which locateSpace or locateViews will return values that respect this change.

poseValid :: Bool

poseValid is true if the runtime can determine the pose of the new space in the previous space before the change.

poseInPreviousSpace :: Posef

poseInPreviousSpace is an Posef defining the position and orientation of the new reference space’s natural origin within the natural reference frame of its previous space.

data EventDataInteractionProfileChanged

XrEventDataInteractionProfileChanged - Notifies the application than the active interaction profile has changed

Member Descriptions

Description

The EventDataInteractionProfileChanged event is sent to the application to notify it that the active input form factor for one or more top level user paths has changed. This event must only be sent for interaction profiles that the application indicated its support for via suggestInteractionProfileBindings. This event must only be sent for running sessions.

The application can call getCurrentInteractionProfile if it wants to change its own behavior based on the active hardware.

Valid Usage (Implicit)

See Also

Session, StructureType, getCurrentInteractionProfile, suggestInteractionProfileBindings

Constructors

EventDataInteractionProfileChanged

Fields

session :: Ptr Session_T

session is the Session for which at least one of the interaction profiles for a top level path has changed.

session must be a valid Session handle

data Offset2Df

XrOffset2Df - Float offset in two dimensions

Member Descriptions

Description

This structure is used for component values that may be fractional (floating-point). If used to represent physical distances, values must be in meters.

See Also

Extent2Df, Rect2Df

Constructors

Offset2Df

Fields

x :: Float

x the floating-point offset in the x direction.

y :: Float

y the floating-point offset in the y direction.

data Extent2Df

XrExtent2Df - Extent in two dimensions

Member Descriptions

Description

This structure is used for component values that may be fractional (floating-point). If used to represent physical distances, values must be in meters.

The width and height value must be non-negative.

See Also

CompositionLayerQuad, Offset2Df, Rect2Df, getReferenceSpaceBoundsRect

Constructors

Extent2Df

Fields

width :: Float

width the floating-point width of the extent.

height :: Float

height the floating-point height of the extent.

data Rect2Df

XrRect2Df - Rect in two dimensions

Member Descriptions

Description

This structure is used for component values that may be fractional (floating-point).

See Also

Extent2Df, Offset2Df

Constructors

Rect2Df

Fields

offset :: Offset2Df

offset is the Offset2Df specifying the rectangle offset.

extent :: Extent2Df

extent is the Extent2Df specifying the rectangle extent.

data BaseInStructure

XrBaseInStructure - Convenience type for iterating (read only)

Member Descriptions

Description

BaseInStructure can be used to facilitate iterating through a read-only structure pointer chain.

See Also

BaseInStructure, BaseOutStructure, StructureType

Constructors

BaseInStructure

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

next :: Ptr BaseInStructure

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR.

data BaseOutStructure

XrBaseOutStructure - Convenience type for iterating (mutable)

Member Descriptions

Description

BaseOutStructure can be used to facilitate iterating through a structure pointer chain that returns data back to the application.

See Also

BaseInStructure, BaseOutStructure, StructureType

Constructors

BaseOutStructure

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

next :: Ptr BaseOutStructure

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR.

newtype ObjectType

XrObjectType - Specify an enumeration to track object handle types

Description

The ObjectType enumeration defines values, each of which corresponds to a specific OpenXR handle type. These values can be used to associate debug information with a particular type of object through one or more extensions.

The following table defines ObjectType and OpenXR Handle relationships:

ObjectType 	 OpenXR Handle Type

OBJECT_TYPE_UNKNOWN 	 Unknown/Undefined Handle

OBJECT_TYPE_INSTANCE 	 Instance

OBJECT_TYPE_SESSION 	 Session

OBJECT_TYPE_SWAPCHAIN 	 Swapchain

OBJECT_TYPE_SPACE 	 Space

OBJECT_TYPE_ACTION_SET 	 ActionSet

OBJECT_TYPE_ACTION 	 Action

See Also

DebugUtilsObjectNameInfoEXT

Constructors

ObjectType Int32

Bundled Patterns

pattern OBJECT_TYPE_UNKNOWN :: ObjectType

pattern OBJECT_TYPE_DEBUG_UTILS_MESSENGER_EXT :: ObjectType

pattern OBJECT_TYPE_INSTANCE :: ObjectType

pattern OBJECT_TYPE_SESSION :: ObjectType

pattern OBJECT_TYPE_SWAPCHAIN :: ObjectType

pattern OBJECT_TYPE_SPACE :: ObjectType

pattern OBJECT_TYPE_ACTION_SET :: ObjectType

pattern OBJECT_TYPE_ACTION :: ObjectType

pattern OBJECT_TYPE_HAND_TRACKER_EXT :: ObjectType

pattern OBJECT_TYPE_SPATIAL_ANCHOR_MSFT :: ObjectType