OpenXR.Core10.ViewConfigurations

Documentation

enumerateViewConfigurations :: forall io. MonadIO io => Instance -> SystemId -> io ("viewConfigurationTypes" ::: Vector ViewConfigurationType)

xrEnumerateViewConfigurations - Enumerates supported view configurations

Parameter Descriptions

instance is the instance from which systemId was retrieved.

systemId is the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId whose view configurations will be enumerated.

viewConfigurationsTypeCapacityInput is the capacity of the viewConfigurations array, or 0 to indicate a request to retrieve the required capacity.

viewConfigurationsTypeCountOutput is a pointer to the count of viewConfigurations written, or a pointer to the required capacity in the case that viewConfigurationsTypeCapacityInput is 0.

viewConfigurationsTypes is a pointer to an array of ViewConfigurationType values, but can be NULL if viewConfigurationsTypeCapacityInput is 0.

See Buffer Size Parameters chapter for a detailed description of retrieving the required viewConfigurations size.

Description

enumerateViewConfigurations enumerates the view configuration types supported by the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId. The supported set for that system must not change during the lifetime of its Instance. The returned list of primary view configurations should be in order from what the runtime considered highest to lowest user preference. Thus the first enumerated view configuration type should be the one the runtime prefers the application to use if possible.

Runtimes must always return identical buffer contents from this enumeration for the given systemId and for the lifetime of the instance.

Valid Usage (Implicit)

instance must be a valid Instance handle

viewConfigurationTypeCountOutput must be a pointer to a uint32_t value

If viewConfigurationTypeCapacityInput is not 0, viewConfigurationTypes must be a pointer to an array of viewConfigurationTypeCapacityInput ViewConfigurationType values

Return Codes

Success

SUCCESS

Failure

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_SIZE_INSUFFICIENT

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, ViewConfigurationType

getViewConfigurationProperties

Arguments

:: forall io. MonadIO io

=> Instance
instance is the instance from which systemId was retrieved.

instance must be a valid Instance handle

-> SystemId
systemId is the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId whose view configuration is being queried.

-> ViewConfigurationType
viewConfigurationType is the ViewConfigurationType of the configuration to get.

viewConfigurationType must be a valid ViewConfigurationType value

-> io ViewConfigurationProperties

xrGetViewConfigurationProperties - Gets information for a view configuration

Parameter Descriptions

Description

getViewConfigurationProperties queries properties of an individual view configuration. Applications must use one of the supported view configuration types returned by enumerateViewConfigurations. If viewConfigurationType is not supported by this Instance the runtime must return ERROR_VIEW_CONFIGURATION_TYPE_UNSUPPORTED.

Return Codes

Success

SUCCESS

Failure

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_VIEW_CONFIGURATION_TYPE_UNSUPPORTED

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, ViewConfigurationProperties, ViewConfigurationType

enumerateViewConfigurationViews

Arguments

:: forall a io. (Extendss ViewConfigurationView a, PokeChain a, PeekChain a, MonadIO io)

=> Instance
instance is the instance from which systemId was retrieved.

-> SystemId
systemId is the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId whose view configuration is being queried.

-> ViewConfigurationType
viewConfigurationType is the ViewConfigurationType of the configuration to get.

-> io ("views" ::: Vector (ViewConfigurationView a))

xrEnumerateViewConfigurationViews - Gets view configuration views

Parameter Descriptions

Description

Each ViewConfigurationType defines the number of views associated with it. Applications can query more details of each view element using enumerateViewConfigurationViews. If the supplied viewConfigurationType is not supported by this Instance and https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, the runtime must return ERROR_VIEW_CONFIGURATION_TYPE_UNSUPPORTED.

Runtimes must always return identical buffer contents from this enumeration for the given systemId and viewConfigurationType for the lifetime of the instance.

Valid Usage (Implicit)

instance must be a valid Instance handle

viewConfigurationType must be a valid ViewConfigurationType value

viewCountOutput must be a pointer to a uint32_t value

If viewCapacityInput is not 0, views must be a pointer to an array of viewCapacityInput ViewConfigurationView structures

Return Codes

Success

SUCCESS

Failure

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_VIEW_CONFIGURATION_TYPE_UNSUPPORTED

ERROR_SIZE_INSUFFICIENT

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, ViewConfigurationType, ViewConfigurationView, getViewConfigurationProperties

data ViewConfigurationView (es :: [Type])

XrViewConfigurationView - Individual view configuration

Member Descriptions

Description

See SwapchainSubImage for more information about imageRect values, and SwapchainCreateInfo for more information about creating swapchains appropriately sized to support those imageRect values.

The array of ViewConfigurationView returned by the runtime must adhere to the rules defined in https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#view_configuration_type, such as the count and association to the left and right eyes.

Valid Usage (Implicit)

See Also

StructureType, ViewConfigurationProperties, ViewConfigurationType, enumerateViewConfigurationViews

Constructors

ViewConfigurationView

Fields

next :: Chain es

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR.

next must be NULL or a valid pointer to the next structure in a structure chain. See also: ViewConfigurationDepthRangeEXT, ViewConfigurationViewFovEPIC

recommendedImageRectWidth :: Word32

recommendedImageRectWidth is the optimal width of imageRect to use when rendering this view into a swapchain.

maxImageRectWidth :: Word32

maxImageRectWidth is the maximum width of imageRect supported when rendering this view into a swapchain.

recommendedImageRectHeight :: Word32

recommendedImageRectHeight is the optimal height of imageRect to use when rendering this view into a swapchain.

maxImageRectHeight :: Word32

maxImageRectHeight is the maximum height of imageRect supported when rendering this view into a swapchain.

recommendedSwapchainSampleCount :: Word32

recommendedSwapchainSampleCount is the recommended number of sub-data element samples to create for each swapchain image that will be rendered into for this view.

maxSwapchainSampleCount :: Word32

maxSwapchainSampleCount is the maximum number of sub-data element samples supported for swapchain images that will be rendered into for this view.

data ViewConfigurationProperties

XrViewConfigurationProperties - Detailed configuration properties for an XrViewConfigurationProperties

Valid Usage (Implicit)

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32, StructureType, ViewConfigurationType, ViewConfigurationView, getViewConfigurationProperties

Constructors

ViewConfigurationProperties

Fields

viewConfigurationType :: ViewConfigurationType

viewConfigurationType is the ViewConfigurationType of the configuration.

viewConfigurationType must be a valid ViewConfigurationType value

fovMutable :: Bool

fovMutable indicates if the view field of view can be modified by the application.