OpenXR.Extensions.XR_EXTX_overlay

Description

Name

XR_EXTX_overlay - instance extension

Specification

See XR_EXTX_overlay in the main specification for complete information.

Registered Extension Number

34

Revision

4

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

EventDataMainSessionVisibilityChangedEXTX, SessionCreateInfoOverlayEXTX

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data SessionCreateInfoOverlayEXTX

XrSessionCreateInfoOverlayEXTX - Session creation extension struct providing overlay session parameters

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SessionCreateInfoOverlayEXTX

type must be TYPE_SESSION_CREATE_INFO_OVERLAY_EXTX

next must be NULL or a valid pointer to the next structure in a structure chain

createFlags must be a valid combination of XrOverlaySessionCreateFlagBitsEXTX values

createFlags must not be 0

See Also

OverlaySessionCreateFlagsEXTX, StructureType

Constructors

SessionCreateInfoOverlayEXTX

Fields

createFlags :: OverlaySessionCreateFlagsEXTX

createFlags is 0 or one or more XrOverlaySessionCreateFlagBitsEXTX which indicate various characteristics desired for the overlay session.

sessionLayersPlacement :: Word32

sessionLayersPlacement is a value indicating the desired placement of the session’s composition layers in terms of other sessions.

data EventDataMainSessionVisibilityChangedEXTX

XrEventDataMainSessionVisibilityChangedEXTX - Event representing main session visibility change

Members

Receiving the EventDataMainSessionVisibilityChangedEXTX event structure indicates that the main session has gained or lost visibility. This can occur in many cases, one typical example is when a user switches from one OpenXR application to another. See EventDataMainSessionVisibilityChangedEXTX for more information on the standard behavior. This structure contains additional information on the main session including flags which indicate additional state information of the main session. Currently, the only flag value supplied is XR_OVERLAY_MAIN_SESSION_ENABLED_COMPOSITION_LAYER_INFO_DEPTH_BIT_EXTX which indicates if the main session has enabled the https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XR_KHR_composition_layer_depth extension.

Member Descriptions

type is the StructureType of this structure.

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR or this extension.

visible is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32 which indicates if session is now visible or is not.

flags is 0 or one or more XrOverlayMainSessionFlagBitsEXTX which indicates various state information for the main session.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using EventDataMainSessionVisibilityChangedEXTX

type must be TYPE_EVENT_DATA_MAIN_SESSION_VISIBILITY_CHANGED_EXTX

next must be NULL or a valid pointer to the next structure in a structure chain

flags must be a valid combination of XrOverlayMainSessionFlagBitsEXTX values

flags must not be 0

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32, OverlayMainSessionFlagsEXTX, StructureType

Constructors

EventDataMainSessionVisibilityChangedEXTX

Fields

visible :: Bool

flags :: OverlayMainSessionFlagsEXTX

newtype OverlayMainSessionFlagsEXTX

XrOverlayMainSessionFlagsEXTX - XrOverlayMainSessionFlagsEXTX

See Also

EventDataMainSessionVisibilityChangedEXTX

Constructors

OverlayMainSessionFlagsEXTX Flags64

newtype OverlaySessionCreateFlagsEXTX

XrOverlaySessionCreateFlagsEXTX - XrOverlaySessionCreateFlagsEXTX

See Also

SessionCreateInfoOverlayEXTX

Constructors

OverlaySessionCreateFlagsEXTX Flags64

type EXTX_overlay_SPEC_VERSION = 4

pattern EXTX_overlay_SPEC_VERSION :: forall a. Integral a => a

type EXTX_OVERLAY_EXTENSION_NAME = "XR_EXTX_overlay"

pattern EXTX_OVERLAY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a