OpenXR.Extensions.XR_EXT_eye_gaze_interaction

Description

Name

XR_EXT_eye_gaze_interaction - instance extension

Specification

See XR_EXT_eye_gaze_interaction in the main specification for complete information.

Registered Extension Number

31

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

EyeGazeSampleTimeEXT, SystemEyeGazeInteractionPropertiesEXT

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data SystemEyeGazeInteractionPropertiesEXT

XrSystemEyeGazeInteractionPropertiesEXT - Eye gaze interaction system properties

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SystemEyeGazeInteractionPropertiesEXT

type must be TYPE_SYSTEM_EYE_GAZE_INTERACTION_PROPERTIES_EXT

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32, StructureType

Constructors

SystemEyeGazeInteractionPropertiesEXT

Fields

supportsEyeGazeInteraction :: Bool

supportsEyeGazeInteraction the runtime must set this value to TRUE when eye gaze sufficient for use cases such as aiming or targeting is supported by the current device, otherwise the runtime must set this to FALSE.

data EyeGazeSampleTimeEXT

XrEyeGazeSampleTimeEXT - Eye gaze sample time structure

Valid Usage (Implicit)

The @@ extension must be enabled prior to using EyeGazeSampleTimeEXT

type must be TYPE_EYE_GAZE_SAMPLE_TIME_EXT

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime

Constructors

EyeGazeSampleTimeEXT

Fields

time :: Time

time is when in time the eye gaze pose is expressed.

type EXT_eye_gaze_interaction_SPEC_VERSION = 1

pattern EXT_eye_gaze_interaction_SPEC_VERSION :: forall a. Integral a => a

type EXT_EYE_GAZE_INTERACTION_EXTENSION_NAME = "XR_EXT_eye_gaze_interaction"

pattern EXT_EYE_GAZE_INTERACTION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a