OpenXR.Extensions.XR_EXT_hand_tracking

Description

Name

XR_EXT_hand_tracking - instance extension

Specification

See XR_EXT_hand_tracking in the main specification for complete information.

Registered Extension Number

52

Revision

2

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

HAND_JOINT_COUNT_EXT, HandEXT, HandJointEXT, HandJointLocationEXT, HandJointLocationsEXT, HandJointSetEXT, HandJointVelocitiesEXT, HandJointVelocityEXT, HandJointsLocateInfoEXT, HandTrackerCreateInfoEXT, SystemHandTrackingPropertiesEXT, createHandTrackerEXT, destroyHandTrackerEXT, locateHandJointsEXT

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

createHandTrackerEXT

Arguments

:: forall a io. (Extendss HandTrackerCreateInfoEXT a, PokeChain a, MonadIO io)

=> Session
session is an Session in which the hand tracker will be active.

-> HandTrackerCreateInfoEXT a
createInfo is the HandTrackerCreateInfoEXT used to specify the hand tracker.

-> io (Result, HandTrackerEXT)

xrCreateHandTrackerEXT - Create a hand joints handle.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling createHandTrackerEXT

session must be a valid Session handle

createInfo must be a pointer to a valid HandTrackerCreateInfoEXT structure

handTracker must be a pointer to an HandTrackerEXT handle

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

ERROR_FEATURE_UNSUPPORTED

If the system does not support hand tracking, runtime must return ERROR_FEATURE_UNSUPPORTED from createHandTrackerEXT. In this case, the runtime must return FALSE for supportsHandTracking in SystemHandTrackingPropertiesEXT when the function getSystemProperties is called, so that the application can avoid creating a hand tracker.

See Also

HandTrackerCreateInfoEXT, HandTrackerEXT, Session

withHandTrackerEXT :: forall a io r. (Extendss HandTrackerCreateInfoEXT a, PokeChain a, MonadIO io) => Session -> HandTrackerCreateInfoEXT a -> (io (Result, HandTrackerEXT) -> ((Result, HandTrackerEXT) -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createHandTrackerEXT and destroyHandTrackerEXT

To ensure that destroyHandTrackerEXT is always called: pass bracket (or the allocate function from your favourite resource management library) as the last argument. To just extract the pair pass (,) as the last argument.

destroyHandTrackerEXT

Arguments

:: forall io. MonadIO io

=> HandTrackerEXT
handTracker is an HandTrackerEXT previously created by createHandTrackerEXT.

-> io ()

xrDestroyHandTrackerEXT - Destroy a hand joints handle

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling destroyHandTrackerEXT

handTracker must be a valid HandTrackerEXT handle

Thread Safety

Access to handTracker, and any child handles, must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_FUNCTION_UNSUPPORTED

See Also

HandTrackerEXT

locateHandJointsEXT

Arguments

:: forall a io. (Extendss HandJointLocationsEXT a, PokeChain a, PeekChain a, MonadIO io)

=> HandTrackerEXT
handTracker is an HandTrackerEXT previously created by createHandTrackerEXT.

-> HandJointsLocateInfoEXT
locateInfo is a pointer to HandJointsLocateInfoEXT describing information to locate hand joints.

-> io (Result, HandJointLocationsEXT a)

xrLocateHandJointsEXT - Locate hand joint locations

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling locateHandJointsEXT

handTracker must be a valid HandTrackerEXT handle

locateInfo must be a pointer to a valid HandJointsLocateInfoEXT structure

locations must be a pointer to an HandJointLocationsEXT structure

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_VALIDATION_FAILURE

ERROR_TIME_INVALID

ERROR_FUNCTION_UNSUPPORTED

See Also

HandJointLocationsEXT, HandJointsLocateInfoEXT, HandTrackerEXT

data SystemHandTrackingPropertiesEXT

XrSystemHandTrackingPropertiesEXT - System property for hand tracking

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SystemHandTrackingPropertiesEXT

type must be TYPE_SYSTEM_HAND_TRACKING_PROPERTIES_EXT

next must be NULL or a valid pointer to the next structure in a structure chain

If a runtime returns FALSE for supportsHandTracking, the runtime must return ERROR_FEATURE_UNSUPPORTED from createHandTrackerEXT.

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32, StructureType

Constructors

SystemHandTrackingPropertiesEXT

Fields

supportsHandTracking :: Bool

supportsHandTracking is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32, indicating if current system is capable of hand tracking input.

data HandTrackerCreateInfoEXT (es :: [Type])

XrHandTrackerCreateInfoEXT - Information to create a hand joints handle

Valid Usage (Implicit)

The @@ extension must be enabled prior to using HandTrackerCreateInfoEXT

type must be TYPE_HAND_TRACKER_CREATE_INFO_EXT

next must be NULL or a valid pointer to the next structure in a structure chain. See also: HandPoseTypeInfoMSFT

hand must be a valid HandEXT value

handJointSet must be a valid HandJointSetEXT value

See Also

HandEXT, HandJointSetEXT, StructureType, createHandTrackerEXT

Constructors

HandTrackerCreateInfoEXT

Fields

next :: Chain es

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR or this extension.

hand :: HandEXT

hand is an HandEXT which describes which hand the tracker is tracking.

handJointSet :: HandJointSetEXT

handJointSet is an HandJointSetEXT describe the set of hand joints to retrieve.

data HandJointsLocateInfoEXT

XrHandJointsLocateInfoEXT - Describes the information to locate hand joints

Valid Usage (Implicit)

The @@ extension must be enabled prior to using HandJointsLocateInfoEXT

type must be TYPE_HAND_JOINTS_LOCATE_INFO_EXT

next must be NULL or a valid pointer to the next structure in a structure chain

baseSpace must be a valid Space handle

See Also

Space, StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime, locateHandJointsEXT

Constructors

HandJointsLocateInfoEXT

Fields

baseSpace :: Ptr Space_T

baseSpace is an Space within which the returned hand joint locations will be represented.

time :: Time

time is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime at which to locate the hand joints.

data HandJointLocationEXT

XrHandJointLocationEXT - Describes the location and radius of a hand joint

Member Descriptions

Description

If the returned locationFlags has XR_SPACE_LOCATION_POSITION_VALID_BIT set, the returned radius must be a positive value.

If the returned locationFlags has XR_SPACE_LOCATION_POSITION_VALID_BIT unset, the returned radius value is undefined and should be avoided.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using HandJointLocationEXT

locationFlags must be a valid combination of XrSpaceLocationFlagBits values

locationFlags must not be 0

See Also

HandJointLocationsEXT, Posef, SpaceLocationFlags

Constructors

HandJointLocationEXT

Fields

locationFlags :: SpaceLocationFlags

locationFlags is a bitfield, with bit masks defined in XrSpaceLocationFlagBits, to indicate which members contain valid data. If none of the bits are set, no other fields in this structure should be considered to be valid or meaningful.

pose :: Posef

pose is an Posef defining the position and orientation of the origin of a hand joint within the reference frame of the corresponding HandJointsLocateInfoEXT::baseSpace.

radius :: Float

radius is a float value radius of the corresponding joint in units of meters.

data HandJointVelocityEXT

XrHandJointVelocityEXT - Describes the velocity of a hand joint

Valid Usage (Implicit)

The @@ extension must be enabled prior to using HandJointVelocityEXT

velocityFlags must be a valid combination of XrSpaceVelocityFlagBits values

velocityFlags must not be 0

See Also

HandJointVelocitiesEXT, SpaceVelocityFlags, Vector3f

Constructors

HandJointVelocityEXT

Fields

velocityFlags :: SpaceVelocityFlags

velocityFlags is a bitfield, with bit masks defined in XrSpaceVelocityFlagBits, to indicate which members contain valid data. If none of the bits are set, no other fields in this structure should be considered to be valid or meaningful.

linearVelocity :: Vector3f

linearVelocity is the relative linear velocity of the hand joint with respect to and expressed in the reference frame of the corresponding HandJointsLocateInfoEXT::baseSpace, in units of meters per second.

angularVelocity :: Vector3f

angularVelocity is the relative angular velocity of the hand joint with respect to the corresponding HandJointsLocateInfoEXT::baseSpace. The vector’s direction is expressed in the reference frame of the corresponding HandJointsLocateInfoEXT::baseSpace and is parallel to the rotational axis of the hand joint. The vector’s magnitude is the relative angular speed of the hand joint in radians per second. The vector follows the right-hand rule for torque/rotation.

data HandJointLocationsEXT (es :: [Type])

XrHandJointLocationsEXT - Returns the hand joint locations

Member Descriptions

Description

The application must allocate the memory for the output array jointLocations that can contain at least jointCount of HandJointLocationEXT.

The application must set jointCount as described by the HandJointSetEXT when creating the HandTrackerEXT otherwise the runtime must return ERROR_VALIDATION_FAILURE.

The runtime must update the jointLocations array ordered so that the application can index elements using the corresponding hand joint enum (e.g. HandJointEXT) as described by HandJointSetEXT when creating the HandTrackerEXT. For example, when the HandTrackerEXT is created with HAND_JOINT_SET_DEFAULT_EXT, the application must set the jointCount to HAND_JOINT_COUNT_EXT, and the runtime must fill the jointLocations array ordered so that it may be indexed by the HandJointEXT enum.

If the returned isActive is true, the runtime must return all joint locations with both XR_SPACE_LOCATION_POSITION_VALID_BIT and XR_SPACE_LOCATION_ORIENTATION_VALID_BIT set. Although, in this case, some joint space locations may be untracked (i.e. XR_SPACE_LOCATION_POSITION_TRACKED_BIT or XR_SPACE_LOCATION_ORIENTATION_TRACKED_BIT is unset).

If the returned isActive is false, it indicates the hand tracker did not detect the hand input or the application lost input focus. In this case, the runtime must return all jointLocations with neither XR_SPACE_LOCATION_POSITION_VALID_BIT nor XR_SPACE_LOCATION_ORIENTATION_VALID_BIT set.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using HandJointLocationsEXT

type must be TYPE_HAND_JOINT_LOCATIONS_EXT

next must be NULL or a valid pointer to the next structure in a structure chain. See also: HandJointVelocitiesEXT

jointLocations must be a pointer to an array of jointCount HandJointLocationEXT structures

The jointCount parameter must be greater than 0

See Also

https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32, HandJointLocationEXT, StructureType, locateHandJointsEXT

Constructors

HandJointLocationsEXT

Fields

next :: Chain es

next is NULL or a pointer to the next structure in a structure chain, such as HandJointVelocitiesEXT.

isActive :: Bool

isActive is a https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrBool32 indicating if the hand tracker is actively tracking.

jointCount :: Word32

jointCount is a uint32_t describing the count of elements in jointLocations array.

jointLocations :: Ptr HandJointLocationEXT

jointLocations is an array of HandJointLocationEXT receiving the returned hand joint locations.

data HandJointVelocitiesEXT

XrHandJointVelocitiesEXT - Returns the hand joint velocities

Member Descriptions

Description

The application must allocate the memory for the output array jointVelocities that can contain at least jointCount of HandJointVelocityEXT.

The application must input jointCount as described by the HandJointSetEXT when creating the HandTrackerEXT. Otherwise, the runtime must return ERROR_VALIDATION_FAILURE.

The runtime must update the jointVelocities array in the order so that the application can index elements using the corresponding hand joint enum (e.g. HandJointEXT) as described by the HandJointSetEXT when creating the HandTrackerEXT. For example, when the HandTrackerEXT is created with HAND_JOINT_SET_DEFAULT_EXT, the application must set the jointCount to HAND_JOINT_COUNT_EXT, and the returned jointVelocities array must be ordered to be indexed by enum HandJointEXT enum.

If the returned HandJointLocationsEXT::isActive is false, it indicates the hand tracker did not detect a hand input or the application lost input focus. In this case, the runtime must return all jointVelocities with neither XR_SPACE_VELOCITY_LINEAR_VALID_BIT nor XR_SPACE_VELOCITY_ANGULAR_VALID_BIT set.

If an HandJointVelocitiesEXT structure is chained to HandJointLocationsEXT::next, the returned HandJointLocationsEXT::isActive is true, and the velocity is observed or can be calculated by the runtime, the runtime must fill in the linear velocity of each hand joint within the reference frame of baseSpace and set the XR_SPACE_VELOCITY_LINEAR_VALID_BIT. Similarly, if an HandJointVelocitiesEXT structure is chained to HandJointLocationsEXT::next, the returned HandJointLocationsEXT::isActive is true, and the angular velocity is observed or can be calculated by the runtime, the runtime must fill in the angular velocity of each joint within the reference frame of baseSpace and set the XR_SPACE_VELOCITY_ANGULAR_VALID_BIT.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using HandJointVelocitiesEXT

type must be TYPE_HAND_JOINT_VELOCITIES_EXT

next must be NULL or a valid pointer to the next structure in a structure chain

jointVelocities must be a pointer to an array of jointCount HandJointVelocityEXT structures

The jointCount parameter must be greater than 0

See Also

HandJointLocationsEXT, HandJointVelocityEXT, StructureType

Constructors

HandJointVelocitiesEXT

Fields

jointCount :: Word32

jointCount is a uint32_t describing the number of elements in jointVelocities array.

jointVelocities :: Ptr HandJointVelocityEXT

jointVelocities is an array of HandJointVelocityEXT receiving the returned hand joint velocities.

newtype HandEXT

XrHandEXT - Describes which hand the tracker is tracking.

Enumerant Descriptions

See Also

HandTrackerCreateInfoEXT

Constructors

HandEXT Int32

Bundled Patterns

pattern HAND_LEFT_EXT :: HandEXT
HAND_LEFT_EXT specifies the hand tracker will be tracking the user’s left hand.

pattern HAND_RIGHT_EXT :: HandEXT
HAND_RIGHT_EXT specifies the hand tracker will be tracking the user’s right hand.

newtype HandJointEXT

XrHandJointEXT - The name of hand joints that can be tracked

See Also

No cross-references are available

Constructors

HandJointEXT Int32

Bundled Patterns

pattern HAND_JOINT_PALM_EXT :: HandJointEXT

pattern HAND_JOINT_WRIST_EXT :: HandJointEXT

pattern HAND_JOINT_THUMB_METACARPAL_EXT :: HandJointEXT

pattern HAND_JOINT_THUMB_PROXIMAL_EXT :: HandJointEXT

pattern HAND_JOINT_THUMB_DISTAL_EXT :: HandJointEXT

pattern HAND_JOINT_THUMB_TIP_EXT :: HandJointEXT

pattern HAND_JOINT_INDEX_METACARPAL_EXT :: HandJointEXT

pattern HAND_JOINT_INDEX_PROXIMAL_EXT :: HandJointEXT

pattern HAND_JOINT_INDEX_INTERMEDIATE_EXT :: HandJointEXT

pattern HAND_JOINT_INDEX_DISTAL_EXT :: HandJointEXT

pattern HAND_JOINT_INDEX_TIP_EXT :: HandJointEXT

pattern HAND_JOINT_MIDDLE_METACARPAL_EXT :: HandJointEXT

pattern HAND_JOINT_MIDDLE_PROXIMAL_EXT :: HandJointEXT

pattern HAND_JOINT_MIDDLE_INTERMEDIATE_EXT :: HandJointEXT

pattern HAND_JOINT_MIDDLE_DISTAL_EXT :: HandJointEXT

pattern HAND_JOINT_MIDDLE_TIP_EXT :: HandJointEXT

pattern HAND_JOINT_RING_METACARPAL_EXT :: HandJointEXT

pattern HAND_JOINT_RING_PROXIMAL_EXT :: HandJointEXT

pattern HAND_JOINT_RING_INTERMEDIATE_EXT :: HandJointEXT

pattern HAND_JOINT_RING_DISTAL_EXT :: HandJointEXT

pattern HAND_JOINT_RING_TIP_EXT :: HandJointEXT

pattern HAND_JOINT_LITTLE_METACARPAL_EXT :: HandJointEXT

pattern HAND_JOINT_LITTLE_PROXIMAL_EXT :: HandJointEXT

pattern HAND_JOINT_LITTLE_INTERMEDIATE_EXT :: HandJointEXT

pattern HAND_JOINT_LITTLE_DISTAL_EXT :: HandJointEXT

pattern HAND_JOINT_LITTLE_TIP_EXT :: HandJointEXT

newtype HandJointSetEXT

XrHandJointSetEXT - The set of hand joints to track.

Enumerant Descriptions

See Also

HandTrackerCreateInfoEXT

Constructors

HandJointSetEXT Int32

Bundled Patterns

pattern HAND_JOINT_SET_DEFAULT_EXT :: HandJointSetEXT
HAND_JOINT_SET_DEFAULT_EXT indicates that the created HandTrackerEXT tracks the set of hand joints described by HandJointEXT enum, i.e. the locateHandJointsEXT function returns an array of joint locations with the count of HAND_JOINT_COUNT_EXT and can be indexed using HandJointEXT.

type EXT_hand_tracking_SPEC_VERSION = 2

pattern EXT_hand_tracking_SPEC_VERSION :: forall a. Integral a => a

type EXT_HAND_TRACKING_EXTENSION_NAME = "XR_EXT_hand_tracking"

pattern EXT_HAND_TRACKING_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

data HandTrackerEXT

XrHandTrackerEXT - Represents a tracker for a hand joints.

Description

An application creates separate HandTrackerEXT handles for left and right hands. This handle can be used to locate hand joints using locateHandJointsEXT function.

A hand tracker provides accurate fidelity to the user’s actual hand shape. When the hand tracking input requires the user to be holding a controller in their hand, the hand tracking input will appear as the user holding the controller. This input can be used to render the hand shape together with rendering the controller in the hand.

See Also

createHandMeshSpaceMSFT, createHandTrackerEXT, destroyHandTrackerEXT, locateHandJointsEXT, updateHandMeshMSFT

Constructors

HandTrackerEXT

Fields

handTrackerEXTHandle :: Ptr HandTrackerEXT_T

instanceCmds :: InstanceCmds

type HAND_JOINT_COUNT_EXT = 26

pattern HAND_JOINT_COUNT_EXT :: forall a. Integral a => a

XR_HAND_JOINT_COUNT_EXT - The number of hand joint enums defined in XrHandJointEXT

Description

HAND_JOINT_COUNT_EXT defines the number of hand joint enumerants defined in HandJointEXT

See Also

HandJointEXT