OpenXR.Extensions.XR_EXT_performance_settings

Description

Name

XR_EXT_performance_settings - instance extension

Specification

See XR_EXT_performance_settings in the main specification for complete information.

Registered Extension Number

16

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

EventDataPerfSettingsEXT, PerfSettingsDomainEXT, PerfSettingsLevelEXT, PerfSettingsNotificationLevelEXT, PerfSettingsSubDomainEXT, perfSettingsSetPerformanceLevelEXT

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

perfSettingsSetPerformanceLevelEXT

Arguments

:: forall io. MonadIO io

=> Session
session is a valid Session handle.

-> PerfSettingsDomainEXT
domain: the processing domain for which the level hint is applied

-> PerfSettingsLevelEXT
level: the level hint to be applied

-> io Result

xrPerfSettingsSetPerformanceLevelEXT - xrPerfSettingsSetPerformanceLevelEXT

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling perfSettingsSetPerformanceLevelEXT

session must be a valid Session handle

domain must be a valid PerfSettingsDomainEXT value

level must be a valid PerfSettingsLevelEXT value

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_SESSION_LOST

ERROR_HANDLE_INVALID

ERROR_VALIDATION_FAILURE

ERROR_RUNTIME_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

PerfSettingsDomainEXT, PerfSettingsLevelEXT, Session

data EventDataPerfSettingsEXT

XrEventDataPerfSettingsEXT - XrEventDataPerfSettingsEXT

Valid Usage (Implicit)

The @@ extension must be enabled prior to using EventDataPerfSettingsEXT

type must be TYPE_EVENT_DATA_PERF_SETTINGS_EXT

next must be NULL or a valid pointer to the next structure in a structure chain

domain must be a valid PerfSettingsDomainEXT value

subDomain must be a valid PerfSettingsSubDomainEXT value

fromLevel must be a valid PerfSettingsNotificationLevelEXT value

toLevel must be a valid PerfSettingsNotificationLevelEXT value

See Also

PerfSettingsDomainEXT, PerfSettingsNotificationLevelEXT, PerfSettingsSubDomainEXT, StructureType

Constructors

EventDataPerfSettingsEXT

Fields

domain :: PerfSettingsDomainEXT

domain : processing domain in which a threshold has been crossed

subDomain :: PerfSettingsSubDomainEXT

subDomain : system area in which a threshold has been crossed

fromLevel :: PerfSettingsNotificationLevelEXT

fromLevel : enumerated notification level which has been exited

toLevel :: PerfSettingsNotificationLevelEXT

toLevel : enumerated notification level which has been entered

newtype PerfSettingsLevelEXT

XrPerfSettingsLevelEXT - XrPerfSettingsLevelEXT

See Also

perfSettingsSetPerformanceLevelEXT

Constructors

PerfSettingsLevelEXT Int32

Bundled Patterns

pattern PERF_SETTINGS_LEVEL_POWER_SAVINGS_EXT :: PerfSettingsLevelEXT

pattern PERF_SETTINGS_LEVEL_SUSTAINED_LOW_EXT :: PerfSettingsLevelEXT

pattern PERF_SETTINGS_LEVEL_SUSTAINED_HIGH_EXT :: PerfSettingsLevelEXT

pattern PERF_SETTINGS_LEVEL_BOOST_EXT :: PerfSettingsLevelEXT

newtype PerfSettingsDomainEXT

XrPerfSettingsDomainEXT - XrPerfSettingsDomainEXT

See Also

EventDataPerfSettingsEXT, perfSettingsSetPerformanceLevelEXT, thermalGetTemperatureTrendEXT

Constructors

PerfSettingsDomainEXT Int32

Bundled Patterns

pattern PERF_SETTINGS_DOMAIN_CPU_EXT :: PerfSettingsDomainEXT

pattern PERF_SETTINGS_DOMAIN_GPU_EXT :: PerfSettingsDomainEXT

newtype PerfSettingsSubDomainEXT

XrPerfSettingsSubDomainEXT - XrPerfSettingsSubDomainEXT

See Also

EventDataPerfSettingsEXT

Constructors

PerfSettingsSubDomainEXT Int32

Bundled Patterns

pattern PERF_SETTINGS_SUB_DOMAIN_COMPOSITING_EXT :: PerfSettingsSubDomainEXT

pattern PERF_SETTINGS_SUB_DOMAIN_RENDERING_EXT :: PerfSettingsSubDomainEXT

pattern PERF_SETTINGS_SUB_DOMAIN_THERMAL_EXT :: PerfSettingsSubDomainEXT

newtype PerfSettingsNotificationLevelEXT

XrPerfSettingsNotificationLevelEXT - XrPerfSettingsNotificationLevelEXT

See Also

EventDataPerfSettingsEXT, thermalGetTemperatureTrendEXT

Constructors

PerfSettingsNotificationLevelEXT Int32

Bundled Patterns

pattern PERF_SETTINGS_NOTIF_LEVEL_NORMAL_EXT :: PerfSettingsNotificationLevelEXT

pattern PERF_SETTINGS_NOTIF_LEVEL_WARNING_EXT :: PerfSettingsNotificationLevelEXT

pattern PERF_SETTINGS_NOTIF_LEVEL_IMPAIRED_EXT :: PerfSettingsNotificationLevelEXT

type EXT_performance_settings_SPEC_VERSION = 1

pattern EXT_performance_settings_SPEC_VERSION :: forall a. Integral a => a

type EXT_PERFORMANCE_SETTINGS_EXTENSION_NAME = "XR_EXT_performance_settings"

pattern EXT_PERFORMANCE_SETTINGS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a