OpenXR.Extensions.XR_EXT_thermal_query

Description

Name

XR_EXT_thermal_query - instance extension

Specification

See XR_EXT_thermal_query in the main specification for complete information.

Registered Extension Number

17

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

PerfSettingsDomainEXT, PerfSettingsNotificationLevelEXT, thermalGetTemperatureTrendEXT

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

thermalGetTemperatureTrendEXT

Arguments

:: forall io. MonadIO io

=> Session
session is a valid Session handle.

-> PerfSettingsDomainEXT
domain : the processing domain

-> io (Result, PerfSettingsNotificationLevelEXT, "tempHeadroom" ::: Float, "tempSlope" ::: Float)

xrThermalGetTemperatureTrendEXT - xrThermalGetTemperatureTrendEXT

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling thermalGetTemperatureTrendEXT

session must be a valid Session handle

domain must be a valid PerfSettingsDomainEXT value

notificationLevel must be a pointer to an PerfSettingsNotificationLevelEXT value

tempHeadroom must be a pointer to a float value

tempSlope must be a pointer to a float value

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_SESSION_LOST

ERROR_HANDLE_INVALID

ERROR_VALIDATION_FAILURE

ERROR_RUNTIME_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

PerfSettingsDomainEXT, PerfSettingsNotificationLevelEXT, Session

type EXT_thermal_query_SPEC_VERSION = 1

pattern EXT_thermal_query_SPEC_VERSION :: forall a. Integral a => a

type EXT_THERMAL_QUERY_EXTENSION_NAME = "XR_EXT_thermal_query"

pattern EXT_THERMAL_QUERY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype PerfSettingsDomainEXT

XrPerfSettingsDomainEXT - XrPerfSettingsDomainEXT

See Also

EventDataPerfSettingsEXT, perfSettingsSetPerformanceLevelEXT, thermalGetTemperatureTrendEXT

Constructors

PerfSettingsDomainEXT Int32

Bundled Patterns

pattern PERF_SETTINGS_DOMAIN_CPU_EXT :: PerfSettingsDomainEXT

pattern PERF_SETTINGS_DOMAIN_GPU_EXT :: PerfSettingsDomainEXT

newtype PerfSettingsNotificationLevelEXT

XrPerfSettingsNotificationLevelEXT - XrPerfSettingsNotificationLevelEXT

See Also

EventDataPerfSettingsEXT, thermalGetTemperatureTrendEXT

Constructors

PerfSettingsNotificationLevelEXT Int32

Bundled Patterns

pattern PERF_SETTINGS_NOTIF_LEVEL_NORMAL_EXT :: PerfSettingsNotificationLevelEXT

pattern PERF_SETTINGS_NOTIF_LEVEL_WARNING_EXT :: PerfSettingsNotificationLevelEXT

pattern PERF_SETTINGS_NOTIF_LEVEL_IMPAIRED_EXT :: PerfSettingsNotificationLevelEXT