OpenXR.Extensions.XR_FB_color_space

Description

Name

XR_FB_color_space - instance extension

Specification

See XR_FB_color_space in the main specification for complete information.

Registered Extension Number

109

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

ColorSpaceFB, SystemColorSpacePropertiesFB, enumerateColorSpacesFB, setColorSpaceFB

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

enumerateColorSpacesFB :: forall io. MonadIO io => Session -> io (Result, "colorSpaces" ::: Vector ColorSpaceFB)

xrEnumerateColorSpacesFB - Enumerates color spaces

Parameter Descriptions

session is the session that enumerates the supported color spaces.

colorSpaceCapacityInput is the capacity of the colorSpaces array, or 0 to retrieve the required capacity.

colorSpaceCountOutput is a pointer to the count of ColorSpaceFB colorSpaces written, or a pointer to the required capacity in the case that colorSpaceCapacityInput is 0.

colorSpaces is a pointer to an array of ColorSpaceFB color spaces, but can be NULL if colorSpaceCapacityInput is 0.

See Buffer Size Parameters chapter for a detailed description of retrieving the required colorSpaces size.

Description

enumerateColorSpacesFB enumerates the color spaces supported by the current session. Runtimes must always return identical buffer contents from this enumeration for the lifetime of the session.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling enumerateColorSpacesFB

session must be a valid Session handle

colorSpaceCountOutput must be a pointer to a uint32_t value

If colorSpaceCapacityInput is not 0, colorSpaces must be a pointer to an array of colorSpaceCapacityInput ColorSpaceFB values

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SIZE_INSUFFICIENT

ERROR_FUNCTION_UNSUPPORTED

ERROR_VALIDATION_FAILURE

See Also

ColorSpaceFB, Session, setColorSpaceFB

setColorSpaceFB

Arguments

:: forall io. MonadIO io

=> Session
session is a valid Session handle.

-> ColorSpaceFB

-> io Result

xrSetColorSpaceFB - Set a color space

Parameter Descriptions

Description

setColorSpaceFB provides a mechanism for an application to specify the color space used in the final rendered frame. If this function is not called, the session will use the color space deemed appropriate by the runtime. Facebook HMDs for both PC and Mobile product lines default to COLOR_SPACE_RIFT_CV1_FB. The runtime must return ERROR_COLOR_SPACE_UNSUPPORTED_FB if colorSpace is not one of the values enumerated by enumerateColorSpacesFB.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling setColorSpaceFB

session must be a valid Session handle

colorspace must be a valid ColorSpaceFB value

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_COLOR_SPACE_UNSUPPORTED_FB

ERROR_FEATURE_UNSUPPORTED

ERROR_FUNCTION_UNSUPPORTED

ERROR_VALIDATION_FAILURE

See Also

ColorSpaceFB, Session, enumerateColorSpacesFB

data SystemColorSpacePropertiesFB

XrSystemColorSpacePropertiesFB - System property for color space

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SystemColorSpacePropertiesFB

type must be TYPE_SYSTEM_COLOR_SPACE_PROPERTIES_FB

next must be NULL or a valid pointer to the next structure in a structure chain

colorSpace must be a valid ColorSpaceFB value

See Also

ColorSpaceFB, StructureType

Constructors

SystemColorSpacePropertiesFB

Fields

colorSpace :: ColorSpaceFB

colorSpace is the native color space of the XR device.

newtype ColorSpaceFB

XrColorSpaceFB - Color Space Type

Enumerant Descriptions

See Also

SystemColorSpacePropertiesFB, enumerateColorSpacesFB, setColorSpaceFB

Constructors

ColorSpaceFB Int32

Bundled Patterns

pattern COLOR_SPACE_UNMANAGED_FB :: ColorSpaceFB
COLOR_SPACE_UNMANAGED_FB. No color correction, not recommended for production use.

pattern COLOR_SPACE_REC2020_FB :: ColorSpaceFB
COLOR_SPACE_REC2020_FB. Standard Rec. 2020 chromacities. This is the preferred color space for standardized color across all Oculus HMDs with D65 white point.

pattern COLOR_SPACE_REC709_FB :: ColorSpaceFB
COLOR_SPACE_REC709_FB. Standard Rec. 709 chromaticities, similar to sRGB.

pattern COLOR_SPACE_RIFT_CV1_FB :: ColorSpaceFB
COLOR_SPACE_RIFT_CV1_FB. Unique color space, between P3 and Adobe RGB using D75 white point.

Color Space Details with Chromacity Primaries in CIE 1931 xy:

Red: (0.666, 0.334)

Green: (0.238, 0.714)

Blue: (0.139, 0.053)

White: (0.298, 0.318)

pattern COLOR_SPACE_RIFT_S_FB :: ColorSpaceFB
COLOR_SPACE_RIFT_S_FB. Unique color space. Similar to Rec 709 using D75.

Color Space Details with Chromacity Primaries in CIE 1931 xy:

Red: (0.640, 0.330)

Green: (0.292, 0.586)

Blue: (0.156, 0.058)

White: (0.298, 0.318)

pattern COLOR_SPACE_QUEST_FB :: ColorSpaceFB
COLOR_SPACE_QUEST_FB. Unique color space. Similar to Rift CV1 using D75 white point

Color Space Details with Chromacity Primaries in CIE 1931 xy:

Red: (0.661, 0.338)

Green: (0.228, 0.718)

Blue: (0.142, 0.042)

White: (0.298, 0.318)

pattern COLOR_SPACE_P3_FB :: ColorSpaceFB
COLOR_SPACE_P3_FB. Similar to DCI-P3, but uses D65 white point instead.

Color Space Details with Chromacity Primaries in CIE 1931 xy:

Red: (0.680, 0.320)

Green: (0.265, 0.690)

Blue: (0.150, 0.060)

White: (0.313, 0.329)

pattern COLOR_SPACE_ADOBE_RGB_FB :: ColorSpaceFB
COLOR_SPACE_ADOBE_RGB_FB. Standard Adobe chromacities.

type FB_color_space_SPEC_VERSION = 1

pattern FB_color_space_SPEC_VERSION :: forall a. Integral a => a

type FB_COLOR_SPACE_EXTENSION_NAME = "XR_FB_color_space"

pattern FB_COLOR_SPACE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a