OpenXR.Extensions.XR_FB_display_refresh_rate

Description

Name

XR_FB_display_refresh_rate - instance extension

Specification

See XR_FB_display_refresh_rate in the main specification for complete information.

Registered Extension Number

102

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

EventDataDisplayRefreshRateChangedFB, enumerateDisplayRefreshRatesFB, getDisplayRefreshRateFB, requestDisplayRefreshRateFB

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

enumerateDisplayRefreshRatesFB :: forall io. MonadIO io => Session -> io (Result, "displayRefreshRates" ::: Vector Float)

xrEnumerateDisplayRefreshRatesFB - Enumerates display refresh rates

Parameter Descriptions

session is the session that enumerates the supported display refresh rates.

displayRefreshRateCapacityInput is the capacity of the displayRefreshRates, or 0 to retrieve the required capacity.

displayRefreshRateCountOutput is a pointer to the count of float displayRefreshRates written, or a pointer to the required capacity in the case that displayRefreshRateCapacityInput is 0.

displayRefreshRates is a pointer to an array of float display refresh rates, but can be NULL if displayRefreshRateCapacityInput is 0.

See Buffer Size Parameters chapter for a detailed description of retrieving the required displayRefreshRates size.

Description

enumerateDisplayRefreshRatesFB enumerates the display refresh rates supported by the current session. Display refresh rates must be in order from lowest to highest supported display refresh rates. Runtimes must always return identical buffer contents from this enumeration for the lifetime of the session.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling enumerateDisplayRefreshRatesFB

session must be a valid Session handle

displayRefreshRateCountOutput must be a pointer to a uint32_t value

If displayRefreshRateCapacityInput is not 0, displayRefreshRates must be a pointer to an array of displayRefreshRateCapacityInput float values

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SIZE_INSUFFICIENT

ERROR_FUNCTION_UNSUPPORTED

ERROR_VALIDATION_FAILURE

See Also

Session, getDisplayRefreshRateFB, requestDisplayRefreshRateFB

getDisplayRefreshRateFB

Arguments

:: forall io. MonadIO io

=> Session
session is the Session to query.

-> io (Result, "displayRefreshRate" ::: Float)

xrGetDisplayRefreshRateFB - Get the current display refresh rate

Parameter Descriptions

Description

getDisplayRefreshRateFB retrieves the current display refresh rate.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getDisplayRefreshRateFB

session must be a valid Session handle

displayRefreshRate must be a pointer to a float value

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_FUNCTION_UNSUPPORTED

ERROR_VALIDATION_FAILURE

See Also

Session, enumerateDisplayRefreshRatesFB, requestDisplayRefreshRateFB

requestDisplayRefreshRateFB

Arguments

:: forall io. MonadIO io

=> Session
session is a valid Session handle.

-> ("displayRefreshRate" ::: Float)
displayRefreshRate is 0.0f or a supported display refresh rate. Supported display refresh rates are indicated by enumerateDisplayRefreshRatesFB.

-> io Result

xrRequestDisplayRefreshRateFB - Request a display refresh rate

Parameter Descriptions

Description

requestDisplayRefreshRateFB provides a mechanism for an application to request the system to dynamically change the display refresh rate to the application preferred value. The runtime must return ERROR_DISPLAY_REFRESH_RATE_UNSUPPORTED_FB if displayRefreshRate is not either 0.0f or one of the values enumerated by enumerateDisplayRefreshRatesFB. A display refresh rate of 0.0f indicates the application has no preference.

Note that this is only a request and does not guarantee the system will switch to the requested display refresh rate.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling requestDisplayRefreshRateFB

session must be a valid Session handle

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_DISPLAY_REFRESH_RATE_UNSUPPORTED_FB

ERROR_FEATURE_UNSUPPORTED

ERROR_FUNCTION_UNSUPPORTED

ERROR_VALIDATION_FAILURE

See Also

Session, enumerateDisplayRefreshRatesFB, getDisplayRefreshRateFB

data EventDataDisplayRefreshRateChangedFB

XrEventDataDisplayRefreshRateChangedFB - Event representing display refresh rate change

Valid Usage (Implicit)

The @@ extension must be enabled prior to using EventDataDisplayRefreshRateChangedFB

type must be TYPE_EVENT_DATA_DISPLAY_REFRESH_RATE_CHANGED_FB

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, getDisplayRefreshRateFB

Constructors

EventDataDisplayRefreshRateChangedFB

Fields

fromDisplayRefreshRate :: Float

fromDisplayRefreshRate is the previous display refresh rate.

toDisplayRefreshRate :: Float

toDisplayRefreshRate is the new display refresh rate.

type FB_display_refresh_rate_SPEC_VERSION = 1

pattern FB_display_refresh_rate_SPEC_VERSION :: forall a. Integral a => a

type FB_DISPLAY_REFRESH_RATE_EXTENSION_NAME = "XR_FB_display_refresh_rate"

pattern FB_DISPLAY_REFRESH_RATE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a