OpenXR.Extensions.XR_KHR_D3D11_enable

Description

Name

XR_KHR_D3D11_enable - instance extension

Specification

See XR_KHR_D3D11_enable in the main specification for complete information.

Registered Extension Number

28

Revision

5

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

GraphicsBindingD3D11KHR, GraphicsRequirementsD3D11KHR, SwapchainImageD3D11KHR, getD3D11GraphicsRequirementsKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

getD3D11GraphicsRequirementsKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> SystemId
systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

-> io GraphicsRequirementsD3D11KHR

xrGetD3D11GraphicsRequirementsKHR - Retrieve the D3D11 feature level and graphics device requirements for an instance and system

Parameter Descriptions

Description

The getD3D11GraphicsRequirementsKHR function identifies to the application what graphics device (Windows LUID) needs to be used and the minimum feature level to use. The runtime must return ERROR_GRAPHICS_REQUIREMENTS_CALL_MISSING (ERROR_VALIDATION_FAILURE may be returned due to legacy behavior) on calls to createSession if getD3D11GraphicsRequirementsKHR has not been called for the same instance and systemId. The LUID and feature level that getD3D11GraphicsRequirementsKHR returns should be used to create the ID3D11Device that the application passes to createSession in the GraphicsBindingD3D11KHR.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getD3D11GraphicsRequirementsKHR

instance must be a valid Instance handle

graphicsRequirements must be a pointer to an GraphicsRequirementsD3D11KHR structure

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

GraphicsRequirementsD3D11KHR, Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, createSession

data GraphicsBindingD3D11KHR

XrGraphicsBindingD3D11KHR - The graphics binding structure to be passed at session creation to use D3D11

Member Descriptions

Description

When creating a D3D11-backed Session, the application will provide a pointer to an GraphicsBindingD3D11KHR in the next chain of the SessionCreateInfo.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingD3D11KHR

type must be TYPE_GRAPHICS_BINDING_D3D11_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

device must be a pointer to an ID3D11Device value

See Also

StructureType, createSession

Constructors

GraphicsBindingD3D11KHR

Fields

device :: Ptr ID3D11Device

device is a pointer to a valid ID3D11Device to use.

data SwapchainImageD3D11KHR

XrSwapchainImageD3D11KHR - D3D11-specific swapchain image structure

Member Descriptions

Description

If a given session was created with GraphicsBindingD3D11KHR, the following conditions must apply.

Calls to enumerateSwapchainImages on an Swapchain in that session must return an array of SwapchainImageD3D11KHR structures.

Whenever an OpenXR function accepts an SwapchainImageBaseHeader pointer as a parameter in that session, the runtime must also accept a pointer to an SwapchainImageD3D11KHR.

The OpenXR runtime must interpret the top-left corner of the swapchain image as the coordinate origin unless specified otherwise by extension functionality.

The OpenXR runtime must interpret the swapchain images in a clip space of positive Y pointing up, near Z plane at 0, and far Z plane at 1.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SwapchainImageD3D11KHR

type must be TYPE_SWAPCHAIN_IMAGE_D3D11_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

texture must be a pointer to an ID3D11Texture2D value

See Also

StructureType, SwapchainImageBaseHeader

Constructors

SwapchainImageD3D11KHR

Fields

texture :: Ptr ID3D11Texture2D

texture is a pointer to a valid ID3D11Texture2D to use.

data GraphicsRequirementsD3D11KHR

XrGraphicsRequirementsD3D11KHR - D3D11 feature level and LUID requirements

Member Descriptions

Description

GraphicsRequirementsD3D11KHR is populated by getD3D11GraphicsRequirementsKHR.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsRequirementsD3D11KHR

type must be TYPE_GRAPHICS_REQUIREMENTS_D3D11_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

adapterLuid must be a valid LUID value

minFeatureLevel must be a valid D3D_FEATURE_LEVEL value

See Also

StructureType, getD3D11GraphicsRequirementsKHR

Constructors

GraphicsRequirementsD3D11KHR

Fields

adapterLuid :: LUID

adapterLuid identifies what graphics device needs to be used.

minFeatureLevel :: D3D_FEATURE_LEVEL

minFeatureLevel is the minimum feature level that the D3D11 device must be initialized with.

type KHR_D3D11_enable_SPEC_VERSION = 5

pattern KHR_D3D11_enable_SPEC_VERSION :: forall a. Integral a => a

type KHR_D3D11_ENABLE_EXTENSION_NAME = "XR_KHR_D3D11_enable"

pattern KHR_D3D11_ENABLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type LUID = Word64

type D3D_FEATURE_LEVEL = Word32

data ID3D11Device

data ID3D11Texture2D