OpenXR.Extensions.XR_KHR_android_thread_settings

Description

Name

XR_KHR_android_thread_settings - instance extension

Specification

See XR_KHR_android_thread_settings in the main specification for complete information.

Registered Extension Number

4

Revision

5

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

AndroidThreadTypeKHR, setAndroidApplicationThreadKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

setAndroidApplicationThreadKHR

Arguments

:: forall io. MonadIO io

=> Session
session is a valid Session handle.

-> AndroidThreadTypeKHR
threadType is a classification of the declared thread allowing the XR runtime to apply the relevant priority and attributes. If such settings fail, the runtime must return ERROR_ANDROID_THREAD_SETTINGS_FAILURE_KHR.

-> ("threadId" ::: Word32)
threadId is the kernel thread ID of the declared thread, as returned by gettid() or android.os.process.myTid(). If the thread ID is invalid, the runtime must return ERROR_ANDROID_THREAD_SETTINGS_ID_INVALID_KHR.

-> io Result

xrSetAndroidApplicationThreadKHR - declare threads to be of a certain priority type

Parameter Descriptions

Description

setAndroidApplicationThreadKHR allows to declare an XR-critical thread and to classify it.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling setAndroidApplicationThreadKHR

session must be a valid Session handle

threadType must be a valid AndroidThreadTypeKHR value

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_INSTANCE_LOST

ERROR_SESSION_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_ANDROID_THREAD_SETTINGS_ID_INVALID_KHR

ERROR_ANDROID_THREAD_SETTINGS_FAILURE_KHR

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

AndroidThreadTypeKHR, Session

newtype AndroidThreadTypeKHR

XrAndroidThreadTypeKHR - Enum describing Android thread types

Enumerants

See Also

setAndroidApplicationThreadKHR

Constructors

AndroidThreadTypeKHR Int32

Bundled Patterns

pattern ANDROID_THREAD_TYPE_APPLICATION_MAIN_KHR :: AndroidThreadTypeKHR
ANDROID_THREAD_TYPE_APPLICATION_MAIN_KHR hints the XR runtime that the thread is doing background CPU tasks

pattern ANDROID_THREAD_TYPE_APPLICATION_WORKER_KHR :: AndroidThreadTypeKHR
ANDROID_THREAD_TYPE_APPLICATION_WORKER_KHR hints the XR runtime that the thread is doing time critical CPU tasks

pattern ANDROID_THREAD_TYPE_RENDERER_MAIN_KHR :: AndroidThreadTypeKHR
ANDROID_THREAD_TYPE_RENDERER_MAIN_KHR hints the XR runtime that the thread is doing background graphics device tasks

pattern ANDROID_THREAD_TYPE_RENDERER_WORKER_KHR :: AndroidThreadTypeKHR
ANDROID_THREAD_TYPE_RENDERER_WORKER_KHR hints the XR runtime that the thread is doing time critical graphics device tasks

type KHR_android_thread_settings_SPEC_VERSION = 5

pattern KHR_android_thread_settings_SPEC_VERSION :: forall a. Integral a => a

type KHR_ANDROID_THREAD_SETTINGS_EXTENSION_NAME = "XR_KHR_android_thread_settings"

pattern KHR_ANDROID_THREAD_SETTINGS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a