OpenXR.Extensions.XR_KHR_binding_modification

Description

Name

XR_KHR_binding_modification - instance extension

Specification

See XR_KHR_binding_modification in the main specification for complete information.

Registered Extension Number

121

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

BindingModificationBaseHeaderKHR, BindingModificationsKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data BindingModificationsKHR

XrBindingModificationsKHR - Suggested bindings with binding modification details

Valid Usage (Implicit)

The @@ extension must be enabled prior to using BindingModificationsKHR

type must be TYPE_BINDING_MODIFICATIONS_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

If bindingModificationCount is not 0, bindingModifications must be a pointer to an array of bindingModificationCount valid BindingModificationBaseHeaderKHR-based structures

See Also

BindingModificationBaseHeaderKHR, StructureType, suggestInteractionProfileBindings

Constructors

BindingModificationsKHR

Fields

bindingModificationCount :: Word32

bindingModificationCount is the number of binding modifications in the array pointed to by bindingModifications.

bindingModifications :: Vector BindingModificationBaseHeaderKHR

bindingModifications is a pointer to an array of pointers to binding modification structures based on BindingModificationBaseHeaderKHR, that define all of the application’s suggested binding modifications for the specified interaction profile.

data BindingModificationBaseHeaderKHR

XrBindingModificationBaseHeaderKHR - Base struct for all binding modifications

Member Descriptions

Description

The BindingModificationBaseHeaderKHR is a base structure is overridden by XrBindingModification* child structures.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using BindingModificationBaseHeaderKHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

BindingModificationsKHR, StructureType, suggestInteractionProfileBindings

Constructors

BindingModificationBaseHeaderKHR

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

type KHR_binding_modification_SPEC_VERSION = 1

pattern KHR_binding_modification_SPEC_VERSION :: forall a. Integral a => a

type KHR_BINDING_MODIFICATION_EXTENSION_NAME = "XR_KHR_binding_modification"

pattern KHR_BINDING_MODIFICATION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a