OpenXR.Extensions.XR_KHR_composition_layer_cylinder

Description

Name

XR_KHR_composition_layer_cylinder - instance extension

Specification

See XR_KHR_composition_layer_cylinder in the main specification for complete information.

Registered Extension Number

18

Revision

4

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

CompositionLayerCylinderKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data CompositionLayerCylinderKHR

XrCompositionLayerCylinderKHR - Cylindrical layer composition info

Member Descriptions

Description

CompositionLayerCylinderKHR contains the information needed to render a texture onto a cylinder when calling endFrame. CompositionLayerCylinderKHR is an alias type for the base struct CompositionLayerBaseHeader used in FrameEndInfo.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using CompositionLayerCylinderKHR

type must be TYPE_COMPOSITION_LAYER_CYLINDER_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

layerFlags must be 0 or a valid combination of XrCompositionLayerFlagBits values

space must be a valid Space handle

eyeVisibility must be a valid EyeVisibility value

subImage must be a valid SwapchainSubImage structure

See Also

CompositionLayerBaseHeader, CompositionLayerFlags, EyeVisibility, FrameEndInfo, Posef, Space, StructureType, SwapchainSubImage, endFrame

Constructors

CompositionLayerCylinderKHR

Fields

layerFlags :: CompositionLayerFlags

layerFlags specifies options for the layer.

space :: Ptr Space_T

space is the Space in which the pose of the cylinder layer is evaluated over time.

eyeVisibility :: EyeVisibility

subImage :: SwapchainSubImage

subImage identifies the image SwapchainSubImage to use.

pose :: Posef

pose is an Posef defining the position and orientation of the center point of the view of the cylinder within the reference frame of the space.

radius :: Float

radius is the non-negative radius of the cylinder. Values of zero or floating point positive infinity are treated as an infinite cylinder.

centralAngle :: Float

centralAngle is the angle of the visible section of the cylinder, based at 0 radians, in the range of [0, 2π). It grows symmetrically around the 0 radian angle.

aspectRatio :: Float

aspectRatio is the ratio of the visible cylinder section width / height. The height of the cylinder is given by: (cylinder radius × cylinder angle) / aspectRatio.

type KHR_composition_layer_cylinder_SPEC_VERSION = 4

pattern KHR_composition_layer_cylinder_SPEC_VERSION :: forall a. Integral a => a

type KHR_COMPOSITION_LAYER_CYLINDER_EXTENSION_NAME = "XR_KHR_composition_layer_cylinder"

pattern KHR_COMPOSITION_LAYER_CYLINDER_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a