OpenXR.Extensions.XR_KHR_composition_layer_depth

Description

Name

XR_KHR_composition_layer_depth - instance extension

Specification

See XR_KHR_composition_layer_depth in the main specification for complete information.

Registered Extension Number

11

Revision

5

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

CompositionLayerDepthInfoKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data CompositionLayerDepthInfoKHR

XrCompositionLayerDepthInfoKHR - Depth map layer info

Member Descriptions

Description

CompositionLayerDepthInfoKHR contains the information needed to specify an extra layer with depth information. When submitting depth buffers along with projection layers, add the CompositionLayerDepthInfoKHR to the next chain for all CompositionLayerProjectionView structures in the given layer.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using CompositionLayerDepthInfoKHR

type must be TYPE_COMPOSITION_LAYER_DEPTH_INFO_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

subImage must be a valid SwapchainSubImage structure

See Also

CompositionLayerBaseHeader, CompositionLayerProjection, CompositionLayerProjectionView, FrameEndInfo, StructureType, SwapchainSubImage, endFrame

Constructors

CompositionLayerDepthInfoKHR

Fields

subImage :: SwapchainSubImage

subImage identifies the depth image SwapchainSubImage to be associated with the color swapchain. The contained imageRect specifies the valid portion of the depth image to use, in pixels. It also implicitly defines the transform from normalized image coordinates into pixel coordinates. The contained imageArrayIndex is the depth image array index, with 0 meaning the first or only array element.

minDepth :: Float

minDepth and maxDepth are the range of depth values the depthSwapchain could have, in the range of [0.0,1.0]. This is akin to min and max values of OpenGL’s glDepthRange, but with the requirement here that maxDepth ≥ minDepth.

maxDepth :: Float

nearZ :: Float

nearZ is the positive distance in meters of the minDepth value in the depth swapchain. Applications may use a nearZ that is greater than farZ to indicate depth values are reversed. nearZ can be infinite.

farZ :: Float

farZ is the positive distance in meters of the maxDepth value in the depth swapchain. farZ can be infinite. Applications must not use the same value as nearZ.

type KHR_composition_layer_depth_SPEC_VERSION = 5

pattern KHR_composition_layer_depth_SPEC_VERSION :: forall a. Integral a => a

type KHR_COMPOSITION_LAYER_DEPTH_EXTENSION_NAME = "XR_KHR_composition_layer_depth"

pattern KHR_COMPOSITION_LAYER_DEPTH_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a