OpenXR.Extensions.XR_KHR_composition_layer_equirect

Description

Name

XR_KHR_composition_layer_equirect - instance extension

Specification

See XR_KHR_composition_layer_equirect in the main specification for complete information.

Registered Extension Number

19

Revision

3

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

CompositionLayerEquirectKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data CompositionLayerEquirectKHR

XrCompositionLayerEquirectKHR - Equirectangular layer composition info

Member Descriptions

Description

CompositionLayerEquirectKHR contains the information needed to render an equirectangular image onto a sphere when calling endFrame. CompositionLayerEquirectKHR is an alias type for the base struct CompositionLayerBaseHeader used in FrameEndInfo.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using CompositionLayerEquirectKHR

type must be TYPE_COMPOSITION_LAYER_EQUIRECT_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

layerFlags must be 0 or a valid combination of XrCompositionLayerFlagBits values

space must be a valid Space handle

eyeVisibility must be a valid EyeVisibility value

subImage must be a valid SwapchainSubImage structure

See Also

CompositionLayerBaseHeader, CompositionLayerFlags, EyeVisibility, FrameEndInfo, Posef, Space, StructureType, SwapchainSubImage, Vector2f, endFrame

Constructors

CompositionLayerEquirectKHR

Fields

layerFlags :: CompositionLayerFlags

layerFlags specifies options for the layer.

space :: Ptr Space_T

space is the Space in which the pose of the equirect layer is evaluated over time.

eyeVisibility :: EyeVisibility

subImage :: SwapchainSubImage

subImage identifies the image SwapchainSubImage to use.

pose :: Posef

pose is an Posef defining the position and orientation of the center point of the sphere onto which the equirect image data is mapped, relative to the reference frame of the space.

radius :: Float

radius is the non-negative radius of the sphere onto which the equirect image data is mapped. Values of zero or floating point positive infinity are treated as an infinite sphere.

scale :: Vector2f

scale is an Vector2f indicating a scale of the texture coordinates after the mapping to 2D.

bias :: Vector2f

bias is an Vector2f indicating a bias of the texture coordinates after the mapping to 2D.

type KHR_composition_layer_equirect_SPEC_VERSION = 3

pattern KHR_composition_layer_equirect_SPEC_VERSION :: forall a. Integral a => a

type KHR_COMPOSITION_LAYER_EQUIRECT_EXTENSION_NAME = "XR_KHR_composition_layer_equirect"

pattern KHR_COMPOSITION_LAYER_EQUIRECT_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a