OpenXR.Extensions.XR_KHR_convert_timespec_time

Description

Name

XR_KHR_convert_timespec_time - instance extension

Specification

See XR_KHR_convert_timespec_time in the main specification for complete information.

Registered Extension Number

37

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

convertTimeToTimespecTimeKHR, convertTimespecTimeToTimeKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

convertTimeToTimespecTimeKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> Time
time is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime.

-> io ("timespecTime" ::: Timespec)

xrConvertTimeToTimespecTimeKHR - Convert XrTime to timespec monotonic time

Parameter Descriptions

Description

The convertTimeToTimespecTimeKHR function converts an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime to time as if generated by clock_gettime.

If the output unixTime cannot represent the input time, the runtime must return ERROR_TIME_INVALID.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling convertTimeToTimespecTimeKHR

instance must be a valid Instance handle

timespecTime must be a pointer to a timespec value

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_VALIDATION_FAILURE

ERROR_TIME_INVALID

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime

convertTimespecTimeToTimeKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> ("timespecTime" ::: Timespec)

-> io Time

xrConvertTimespecTimeToTimeKHR - Convert timespec monotonic time to XrTime

Parameter Descriptions

Description

The convertTimespecTimeToTimeKHR function converts a time obtained by the clock_gettime function to the equivalent https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime.

If the output time cannot represent the input unixTime, the runtime must return ERROR_TIME_INVALID.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling convertTimespecTimeToTimeKHR

instance must be a valid Instance handle

timespecTime must be a pointer to a valid timespec value

time must be a pointer to an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime value

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_VALIDATION_FAILURE

ERROR_TIME_INVALID

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime

type KHR_convert_timespec_time_SPEC_VERSION = 1

pattern KHR_convert_timespec_time_SPEC_VERSION :: forall a. Integral a => a

type KHR_CONVERT_TIMESPEC_TIME_EXTENSION_NAME = "XR_KHR_convert_timespec_time"

pattern KHR_CONVERT_TIMESPEC_TIME_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

data Timespec

Constructors

Timespec

Fields

tv_sec :: CTime

tv_nsec :: Int64