OpenXR.Extensions.XR_KHR_loader_init

Description

Name

XR_KHR_loader_init - instance extension

Specification

See XR_KHR_loader_init in the main specification for complete information.

Registered Extension Number

89

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

LoaderInitInfoBaseHeaderKHR, initializeLoaderKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

initializeLoaderKHR

Arguments

:: forall a io. (ToCStruct a, MonadIO io)

=> ("loaderInitInfo" ::: a)
loaderInitInfo is a pointer to an LoaderInitInfoBaseHeaderKHR structure, which is a polymorphic type defined by other platform- or implementation-specific extensions.

-> io ()

xrInitializeLoaderKHR - Initializes loader

Parameter Descriptions

Description

See Also

LoaderInitInfoBaseHeaderKHR

data LoaderInitInfoBaseHeaderKHR

XrLoaderInitInfoBaseHeaderKHR - Initializes OpenXR loader

Valid Usage (Implicit)

The @@ extension must be enabled prior to using LoaderInitInfoBaseHeaderKHR

type must be TYPE_LOADER_INIT_INFO_ANDROID_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, initializeLoaderKHR

Constructors

LoaderInitInfoBaseHeaderKHR

Fields

type' :: StructureType

type is the StructureType of this structure. This base structure itself has no associated StructureType value.

class ToCStruct a => IsLoaderInitInfoKHR a where

Methods

toLoaderInitInfoBaseHeaderKHR :: a -> LoaderInitInfoBaseHeaderKHR

type KHR_loader_init_SPEC_VERSION = 1

pattern KHR_loader_init_SPEC_VERSION :: forall a. Integral a => a

type KHR_LOADER_INIT_EXTENSION_NAME = "XR_KHR_loader_init"

pattern KHR_LOADER_INIT_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a