OpenXR.Extensions.XR_KHR_opengl_enable

Description

Name

XR_KHR_opengl_enable - instance extension

Specification

See XR_KHR_opengl_enable in the main specification for complete information.

Registered Extension Number

24

Revision

9

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

GraphicsBindingOpenGLWaylandKHR, GraphicsBindingOpenGLWin32KHR, GraphicsBindingOpenGLXcbKHR, GraphicsBindingOpenGLXlibKHR, GraphicsRequirementsOpenGLKHR, SwapchainImageOpenGLKHR, getOpenGLGraphicsRequirementsKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

getOpenGLGraphicsRequirementsKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> SystemId
systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

-> io GraphicsRequirementsOpenGLKHR

xrGetOpenGLGraphicsRequirementsKHR - Retrieve the OpenGL version requirements for an instance and system

Parameter Descriptions

Description

The getOpenGLGraphicsRequirementsKHR function identifies to the application the minimum OpenGL version requirement and the highest known tested OpenGL version. The runtime must return ERROR_GRAPHICS_REQUIREMENTS_CALL_MISSING (ERROR_VALIDATION_FAILURE may be returned due to legacy behavior) on calls to createSession if getOpenGLGraphicsRequirementsKHR has not been called for the same instance and systemId.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getOpenGLGraphicsRequirementsKHR

instance must be a valid Instance handle

graphicsRequirements must be a pointer to an GraphicsRequirementsOpenGLKHR structure

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

GraphicsRequirementsOpenGLKHR, Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId

data GraphicsBindingOpenGLWin32KHR

XrGraphicsBindingOpenGLWin32KHR - The graphics binding structure to be passed at session creation to use OpenGL on Windows

Member Descriptions

Description

When creating an OpenGL-backed Session on Microsoft Windows, the application will provide a pointer to an GraphicsBindingOpenGLWin32KHR in the next chain of the SessionCreateInfo. As no standardized way exists for OpenGL to create the graphics context on a specific GPU, the runtime must assume that the application uses the operating systems default GPU. If the GPU used by the runtime does not match the GPU on which the OpenGL context of the application got created, createSession must return ERROR_GRAPHICS_DEVICE_INVALID.

The required window system configuration define to expose this structure type is XR_USE_PLATFORM_WIN32.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingOpenGLWin32KHR

type must be TYPE_GRAPHICS_BINDING_OPENGL_WIN32_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

hDC must be a valid HDC value

hGLRC must be a valid HGLRC value

See Also

StructureType, createSession

Constructors

GraphicsBindingOpenGLWin32KHR

Fields

hDC :: HDC

hDC is a valid Windows HW device context handle.

hGLRC :: HGLRC

hGLRC is a valid Windows OpenGL rendering context handle.

data GraphicsBindingOpenGLXlibKHR

XrGraphicsBindingOpenGLXlibKHR - The graphics binding structure to be passed at session creation to use OpenGL on X11 via Xlib

Member Descriptions

Description

When creating an OpenGL-backed Session on any Linux/Unix platform that utilizes X11 and GLX, via the Xlib library, the application will provide a pointer to an GraphicsBindingOpenGLXlibKHR in the next chain of the SessionCreateInfo.

The required window system configuration define to expose this structure type is XR_USE_PLATFORM_XLIB.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingOpenGLXlibKHR

type must be TYPE_GRAPHICS_BINDING_OPENGL_XLIB_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

xDisplay must be a pointer to a Display value

glxFBConfig must be a valid GLXFBConfig value

glxDrawable must be a valid GLXDrawable value

glxContext must be a valid GLXContext value

See Also

StructureType, createSession

Constructors

GraphicsBindingOpenGLXlibKHR

Fields

xDisplay :: Ptr Display

xDisplay is a valid X11 Display.

visualid :: Word32

visualid is a valid X11 visual identifier.

glxFBConfig :: GLXFBConfig

glxFBConfig is a valid X11 OpenGL GLX GLXFBConfig.

glxDrawable :: GLXDrawable

glxDrawable is a valid X11 OpenGL GLX GLXDrawable.

glxContext :: GLXContext

glxContext is a valid X11 OpenGL GLX GLXContext.

data GraphicsBindingOpenGLXcbKHR

XrGraphicsBindingOpenGLXcbKHR - The graphics binding structure to be passed at session creation to use OpenGL on X11 via XCB

Member Descriptions

Description

When creating an OpenGL-backed Session on any Linux/Unix platform that utilizes X11 and GLX, via the Xlib library, the application will provide a pointer to an GraphicsBindingOpenGLXcbKHR in the next chain of the SessionCreateInfo.

The required window system configuration define to expose this structure type is XR_USE_PLATFORM_XCB.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingOpenGLXcbKHR

type must be TYPE_GRAPHICS_BINDING_OPENGL_XCB_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

connection must be a pointer to an xcb_connection_t value

fbconfigid must be a valid xcb_glx_fbconfig_t value

visualid must be a valid xcb_visualid_t value

glxDrawable must be a valid xcb_glx_drawable_t value

glxContext must be a valid xcb_glx_context_t value

See Also

StructureType, createSession

Constructors

GraphicsBindingOpenGLXcbKHR

Fields

connection :: Ptr Xcb_connection_t

connection is a valid xcb_connection_t.

screenNumber :: Word32

screenNumber is an index indicating which screen should be used for rendering.

fbconfigid :: Xcb_glx_fbconfig_t

fbconfigid is a valid XCB OpenGL GLX xcb_glx_fbconfig_t.

visualid :: Xcb_visualid_t

visualid is a valid XCB OpenGL GLX xcb_visualid_t.

glxDrawable :: Xcb_glx_drawable_t

glxDrawable is a valid XCB OpenGL GLX xcb_glx_drawable_t.

glxContext :: Xcb_glx_context_t

glxContext is a valid XCB OpenGL GLX xcb_glx_context_t.

data GraphicsBindingOpenGLWaylandKHR

XrGraphicsBindingOpenGLWaylandKHR - The graphics binding structure to be passed at session creation to use OpenGL on Wayland

Member Descriptions

Description

When creating an OpenGL-backed Session on any Linux/Unix platform that utilizes the Wayland protocol with its compositor, the application will provide a pointer to an GraphicsBindingOpenGLWaylandKHR in the next chain of the SessionCreateInfo.

The required window system configuration define to expose this structure type is XR_USE_PLATFORM_WAYLAND.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingOpenGLWaylandKHR

type must be TYPE_GRAPHICS_BINDING_OPENGL_WAYLAND_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

display must be a pointer to a wl_display value

See Also

StructureType, createSession

Constructors

GraphicsBindingOpenGLWaylandKHR

Fields

display :: Ptr Wl_display

display is a valid Wayland wl_display.

data SwapchainImageOpenGLKHR

XrSwapchainImageOpenGLKHR - OpenGL-specific swapchain image structure

Member Descriptions

Description

If a given session was created with a XrGraphicsBindingOpenGL*KHR, the following conditions must apply.

Calls to enumerateSwapchainImages on an Swapchain in that session must return an array of SwapchainImageOpenGLKHR structures.

Whenever an OpenXR function accepts an SwapchainImageBaseHeader pointer as a parameter in that session, the runtime must also accept a pointer to an SwapchainImageOpenGLKHR.

The OpenXR runtime must interpret the bottom-left corner of the swapchain image as the coordinate origin unless specified otherwise by extension functionality.

The OpenXR runtime must interpret the swapchain images in a clip space of positive Y pointing up, near Z plane at -1, and far Z plane at 1.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SwapchainImageOpenGLKHR

type must be TYPE_SWAPCHAIN_IMAGE_OPENGL_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, SwapchainImageBaseHeader

Constructors

SwapchainImageOpenGLKHR

Fields

image :: Word32

image is the OpenGL texture handle associated with this swapchain image.

data GraphicsRequirementsOpenGLKHR

XrGraphicsRequirementsOpenGLKHR - OpenGL API version requirements

Member Descriptions

Description

GraphicsRequirementsOpenGLKHR is populated by getOpenGLGraphicsRequirementsKHR with the runtime’s OpenGL API version requirements.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsRequirementsOpenGLKHR

type must be TYPE_GRAPHICS_REQUIREMENTS_OPENGL_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrVersion, getOpenGLGraphicsRequirementsKHR

Constructors

GraphicsRequirementsOpenGLKHR

Fields

minApiVersionSupported :: Version

minApiVersionSupported is the minimum version of OpenGL that the runtime supports. Uses MAKE_VERSION on major and minor API version, ignoring any patch version component.

maxApiVersionSupported :: Version

maxApiVersionSupported is the maximum version of OpenGL that the runtime has been tested on and is known to support. Newer OpenGL versions might work if they are compatible. Uses MAKE_VERSION on major and minor API version, ignoring any patch version component.

type KHR_opengl_enable_SPEC_VERSION = 9

pattern KHR_opengl_enable_SPEC_VERSION :: forall a. Integral a => a

type KHR_OPENGL_ENABLE_EXTENSION_NAME = "XR_KHR_opengl_enable"

pattern KHR_OPENGL_ENABLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type HDC = Ptr ()

type HGLRC = Ptr ()

type Display = Ptr ()

type Xcb_visualid_t = Word32

type Xcb_glx_fbconfig_t = Word32

type Xcb_glx_drawable_t = Word32

type Xcb_glx_context_t = Word32

type GLXFBConfig = Ptr ()

type GLXDrawable = Word64

type GLXContext = Ptr ()

data Xcb_connection_t

data Wl_display