OpenXR.Extensions.XR_KHR_opengl_es_enable

Description

Name

XR_KHR_opengl_es_enable - instance extension

Specification

See XR_KHR_opengl_es_enable in the main specification for complete information.

Registered Extension Number

25

Revision

7

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

GraphicsBindingOpenGLESAndroidKHR, GraphicsRequirementsOpenGLESKHR, SwapchainImageOpenGLESKHR, getOpenGLESGraphicsRequirementsKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

getOpenGLESGraphicsRequirementsKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> SystemId
systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

-> io GraphicsRequirementsOpenGLESKHR

xrGetOpenGLESGraphicsRequirementsKHR - Retrieve the OpenGL ES version requirements for an instance and system

Parameter Descriptions

Description

The getOpenGLESGraphicsRequirementsKHR function identifies to the application the minimum OpenGL ES version requirement and the highest known tested OpenGL ES version. The runtime must return ERROR_GRAPHICS_REQUIREMENTS_CALL_MISSING (ERROR_VALIDATION_FAILURE may be returned due to legacy behavior) on calls to createSession if getOpenGLESGraphicsRequirementsKHR has not been called for the same instance and systemId.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getOpenGLESGraphicsRequirementsKHR

instance must be a valid Instance handle

graphicsRequirements must be a pointer to an GraphicsRequirementsOpenGLESKHR structure

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

GraphicsRequirementsOpenGLESKHR, Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId

data GraphicsBindingOpenGLESAndroidKHR

XrGraphicsBindingOpenGLESAndroidKHR - The graphics binding structure to be passed at session creation to use OpenGL ES on Android

Member Descriptions

Description

When creating an OpenGL ES-backed Session on Android, the application will provide a pointer to an GraphicsBindingOpenGLESAndroidKHR structure in the next chain of the SessionCreateInfo.

The required window system configuration define to expose this structure type is XR_USE_PLATFORM_ANDROID.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingOpenGLESAndroidKHR

type must be TYPE_GRAPHICS_BINDING_OPENGL_ES_ANDROID_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

display must be a valid EGLDisplay value

config must be a valid EGLConfig value

context must be a valid EGLContext value

See Also

StructureType, createSession

Constructors

GraphicsBindingOpenGLESAndroidKHR

Fields

display :: EGLDisplay

display is a valid Android OpenGL ES EGLDisplay.

config :: EGLConfig

config is a valid Android OpenGL ES EGLConfig.

context :: EGLContext

context is a valid Android OpenGL ES EGLContext.

data SwapchainImageOpenGLESKHR

XrSwapchainImageOpenGLESKHR - OpenGL ES-specific swapchain image structure

Member Descriptions

Description

If a given session was created with a XrGraphicsBindingOpenGLES*KHR, the following conditions must apply.

Calls to enumerateSwapchainImages on an Swapchain in that session must return an array of SwapchainImageOpenGLESKHR structures.

Whenever an OpenXR function accepts an SwapchainImageBaseHeader pointer as a parameter in that session, the runtime must also accept a pointer to an SwapchainImageOpenGLESKHR structure.

The OpenXR runtime must interpret the bottom-left corner of the swapchain image as the coordinate origin unless specified otherwise by extension functionality.

The OpenXR runtime must interpret the swapchain images in a clip space of positive Y pointing up, near Z plane at -1, and far Z plane at 1.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SwapchainImageOpenGLESKHR

type must be TYPE_SWAPCHAIN_IMAGE_OPENGL_ES_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, SwapchainImageBaseHeader

Constructors

SwapchainImageOpenGLESKHR

Fields

image :: Word32

image is an index indicating the current OpenGL ES swapchain image to use.

data GraphicsRequirementsOpenGLESKHR

XrGraphicsRequirementsOpenGLESKHR - OpenGL ES API version requirements

Member Descriptions

Description

GraphicsRequirementsOpenGLESKHR is populated by getOpenGLESGraphicsRequirementsKHR with the runtime’s OpenGL ES API version requirements.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsRequirementsOpenGLESKHR

type must be TYPE_GRAPHICS_REQUIREMENTS_OPENGL_ES_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrVersion, getOpenGLESGraphicsRequirementsKHR

Constructors

GraphicsRequirementsOpenGLESKHR

Fields

minApiVersionSupported :: Version

minApiVersionSupported is the minimum version of OpenGL ES that the runtime supports. Uses MAKE_VERSION on major and minor API version, ignoring any patch version component.

maxApiVersionSupported :: Version

maxApiVersionSupported is the maximum version of OpenGL ES that the runtime has been tested on and is known to support. Newer OpenGL ES versions might work if they are compatible. Uses MAKE_VERSION on major and minor API version, ignoring any patch version component.

type KHR_opengl_es_enable_SPEC_VERSION = 7

pattern KHR_opengl_es_enable_SPEC_VERSION :: forall a. Integral a => a

type KHR_OPENGL_ES_ENABLE_EXTENSION_NAME = "XR_KHR_opengl_es_enable"

pattern KHR_OPENGL_ES_ENABLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type EGLDisplay = Ptr ()

type EGLConfig = Ptr ()

type EGLContext = Ptr ()