OpenXR.Extensions.XR_KHR_visibility_mask

Description

Name

XR_KHR_visibility_mask - instance extension

Specification

See XR_KHR_visibility_mask in the main specification for complete information.

Registered Extension Number

32

Revision

2

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

EventDataVisibilityMaskChangedKHR, VisibilityMaskKHR, VisibilityMaskTypeKHR, getVisibilityMaskKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

getVisibilityMaskKHR

Arguments

:: forall io. MonadIO io

=> Session
session is an Session handle previously created with createSession.

-> ViewConfigurationType
viewConfigurationType is the view configuration from which to retrieve mask information.

-> ("viewIndex" ::: Word32)
viewIndex is the individual view within the view configuration from which to retrieve mask information.

-> VisibilityMaskTypeKHR
visibilityMaskType is the type of visibility mask requested.

-> io (Result, VisibilityMaskKHR)

xrGetVisibilityMaskKHR - Gets visibility mask

Parameter Descriptions

Description

getVisibilityMaskKHR retrieves the view mask for a given view. This function follows the two-call idiom for filling multiple buffers in a struct. Specifically, if either vertexCapacityInput or indexCapacityInput is 0, the runtime must respond as if both fields were set to 0, returning the vertex count and index count through vertexCountOutput or indexCountOutput respectively. If a view mask for the specified view isn’t available, the returned vertex and index counts must be 0.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getVisibilityMaskKHR

session must be a valid Session handle

viewConfigurationType must be a valid ViewConfigurationType value

visibilityMaskType must be a valid VisibilityMaskTypeKHR value

visibilityMask must be a pointer to an VisibilityMaskKHR structure

Return Codes

Success

SUCCESS

SESSION_LOSS_PENDING

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_VALIDATION_FAILURE

ERROR_VIEW_CONFIGURATION_TYPE_UNSUPPORTED

ERROR_SIZE_INSUFFICIENT

ERROR_SESSION_LOST

ERROR_FUNCTION_UNSUPPORTED

See Also

Session, ViewConfigurationType, VisibilityMaskKHR, VisibilityMaskTypeKHR

data EventDataVisibilityMaskChangedKHR

XrEventDataVisibilityMaskChangedKHR - Visibility Mask

Valid Usage (Implicit)

The @@ extension must be enabled prior to using EventDataVisibilityMaskChangedKHR

type must be TYPE_EVENT_DATA_VISIBILITY_MASK_CHANGED_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

session must be a valid Session handle

viewConfigurationType must be a valid ViewConfigurationType value

See Also

Session, StructureType, ViewConfigurationType

Constructors

EventDataVisibilityMaskChangedKHR

Fields

session :: Ptr Session_T

session is the Session for which the view mask has changed.

viewConfigurationType :: ViewConfigurationType

viewConfigurationType is the view configuration whose mask has changed.

viewIndex :: Word32

viewIndex is the individual view within the view configuration to which the change refers.

data VisibilityMaskKHR

XrVisibilityMaskKHR - Visibility Mask

Member Descriptions

type is the StructureType of this structure.

next is NULL or a pointer to the next structure in a structure chain. No such structures are defined in core OpenXR or this extension.

vertexCapacityInput is the capacity of the vertices array, or 0 to indicate a request to retrieve the required capacity.

vertexCountOutput is filled in by the runtime with the count of vertices written or the required capacity in the case that vertexCapacityInput or indexCapacityInput is 0.

vertices is an array of vertices filled in by the runtime that specifies mask coordinates in the z=-1 plane of the rendered view—​i.e., one meter in front of the view. When rendering the mask for use in a projection layer, these vertices must be transformed by the application’s projection matrix used for the respective CompositionLayerProjectionView.

Description

indexCapacityInput is the capacity of the indices array, or 0 to indicate a request to retrieve the required capacity.

indexCountOutput is filled in by the runtime with the count of indices written or the required capacity in the case that vertexCapacityInput or indexCapacityInput is 0.

indices is an array of indices filled in by the runtime, specifying the indices of the mask geometry in the vertices array.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using VisibilityMaskKHR

type must be TYPE_VISIBILITY_MASK_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

If vertexCapacityInput is not 0, vertices must be a pointer to an array of vertexCapacityInput Vector2f structures

If indexCapacityInput is not 0, indices must be a pointer to an array of indexCapacityInput uint32_t values

See Also

StructureType, Vector2f, getVisibilityMaskKHR

Constructors

VisibilityMaskKHR

Fields

vertexCapacityInput :: Word32

vertexCountOutput :: Word32

vertices :: Ptr Vector2f

indexCapacityInput :: Word32

indexCountOutput :: Word32

indices :: Ptr Word32

newtype VisibilityMaskTypeKHR

XrVisibilityMaskTypeKHR - Visibility Mask Type

Enumerant Descriptions

See Also

getVisibilityMaskKHR

Constructors

VisibilityMaskTypeKHR Int32

Bundled Patterns

pattern VISIBILITY_MASK_TYPE_HIDDEN_TRIANGLE_MESH_KHR :: VisibilityMaskTypeKHR
VISIBILITY_MASK_TYPE_HIDDEN_TRIANGLE_MESH_KHR refers to a two dimensional triangle mesh on the view surface which should not be drawn to by the application. VisibilityMaskKHR refers to a set of triangles identified by vertices and vertex indices. The index count will thus be a multiple of three. The triangle vertices will be returned in counter-clockwise order as viewed from the user perspective.

pattern VISIBILITY_MASK_TYPE_VISIBLE_TRIANGLE_MESH_KHR :: VisibilityMaskTypeKHR
VISIBILITY_MASK_TYPE_VISIBLE_TRIANGLE_MESH_KHR refers to a two dimensional triangle mesh on the view surface which should be drawn to by the application. VisibilityMaskKHR refers to a set of triangles identified by vertices and vertex indices. The index count will thus be a multiple of three. The triangle vertices will be returned in counter-clockwise order as viewed from the user perspective.

pattern VISIBILITY_MASK_TYPE_LINE_LOOP_KHR :: VisibilityMaskTypeKHR
VISIBILITY_MASK_TYPE_LINE_LOOP_KHR refers to a single multi-segmented line loop on the view surface which encompasses the view area which should be drawn by the application. It is the border that exists between the visible and hidden meshes identified by VISIBILITY_MASK_TYPE_HIDDEN_TRIANGLE_MESH_KHR and VISIBILITY_MASK_TYPE_VISIBLE_TRIANGLE_MESH_KHR. The line is counter-clockwise, contiguous, and non-self crossing, with the last point implicitly connecting to the first point. There is one vertex per point, the index count will equal the vertex count, and the indices will refer to the vertices.

type KHR_visibility_mask_SPEC_VERSION = 2

pattern KHR_visibility_mask_SPEC_VERSION :: forall a. Integral a => a

type KHR_VISIBILITY_MASK_EXTENSION_NAME = "XR_KHR_visibility_mask"

pattern KHR_VISIBILITY_MASK_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a