OpenXR.Extensions.XR_KHR_vulkan_enable

Description

Name

XR_KHR_vulkan_enable - instance extension

Specification

See XR_KHR_vulkan_enable in the main specification for complete information.

Registered Extension Number

26

Revision

7

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

GraphicsBindingVulkanKHR, GraphicsRequirementsVulkanKHR, SwapchainImageVulkanKHR, getVulkanDeviceExtensionsKHR, getVulkanGraphicsDeviceKHR, getVulkanGraphicsRequirementsKHR, getVulkanInstanceExtensionsKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

getVulkanInstanceExtensionsKHR :: forall io. MonadIO io => Instance -> SystemId -> io ("buffer" ::: ByteString)

xrGetVulkanInstanceExtensionsKHR - Get list of required Vulkan instance extensions for an OpenXR instance and system

Parameter Descriptions

instance is an Instance handle previously created with createInstance.

systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

bufferCapacityInput is the capacity of the buffer, or 0 to indicate a request to retrieve the required capacity.

bufferCountOutput is a pointer to the count of characters written (including terminating \0), or a pointer to the required capacity in the case that bufferCapacityInput is 0.

buffer is a pointer to an array of characters, but can be NULL if bufferCapacityInput is 0. The format of the output is a single space (ASCII 0x20) delimited string of extension names.

See Buffer Size Parameters chapter for a detailed description of retrieving the required buffer size.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getVulkanInstanceExtensionsKHR

instance must be a valid Instance handle

bufferCountOutput must be a pointer to a uint32_t value

If bufferCapacityInput is not 0, buffer must be a pointer to an array of bufferCapacityInput null-terminated UTF-8 strings

Return Codes

Success

SUCCESS

Failure

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_SIZE_INSUFFICIENT

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, getVulkanDeviceExtensionsKHR

getVulkanDeviceExtensionsKHR :: forall io. MonadIO io => Instance -> SystemId -> io ("buffer" ::: ByteString)

xrGetVulkanDeviceExtensionsKHR - Get list of required Vulkan device extensions for an OpenXR instance and system

Parameter Descriptions

instance is an Instance handle previously created with createInstance.

systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

bufferCapacityInput is the capacity of the buffer, or 0 to indicate a request to retrieve the required capacity.

bufferCountOutput is a pointer to the count of characters written (including terminating \0), or a pointer to the required capacity in the case that bufferCapacityInput is 0.

buffer is a pointer to an array of characters, but can be NULL if bufferCapacityInput is 0. The format of the output is a single space (ASCII 0x20) delimited string of extension names.

See Buffer Size Parameters chapter for a detailed description of retrieving the required buffer size.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getVulkanDeviceExtensionsKHR

instance must be a valid Instance handle

bufferCountOutput must be a pointer to a uint32_t value

If bufferCapacityInput is not 0, buffer must be a pointer to an array of bufferCapacityInput null-terminated UTF-8 strings

Return Codes

Success

SUCCESS

Failure

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_SIZE_INSUFFICIENT

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, getVulkanInstanceExtensionsKHR

getVulkanGraphicsDeviceKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> SystemId
systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

-> ("vkInstance" ::: Ptr Instance_T)
vkInstance is a valid Vulkan VkInstance.

-> io ("vkPhysicalDevice" ::: Ptr PhysicalDevice_T)

xrGetVulkanGraphicsDeviceKHR - Retrieve the Vulkan physical device associated with an OpenXR instance and system

Parameter Descriptions

Description

getVulkanGraphicsDeviceKHR function identifies to the application what graphics device (Vulkan VkPhysicalDevice) needs to be used. getVulkanGraphicsDeviceKHR must be called prior to calling createSession, and the VkPhysicalDevice that getVulkanGraphicsDeviceKHR returns should be passed to createSession in the GraphicsBindingVulkanKHR.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getVulkanGraphicsDeviceKHR

instance must be a valid Instance handle

vkInstance must be a valid VkInstance value

vkPhysicalDevice must be a pointer to a VkPhysicalDevice value

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId

getVulkanGraphicsRequirementsKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> SystemId
systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

-> io GraphicsRequirementsVulkanKHR

xrGetVulkanGraphicsRequirementsKHR - Retrieve the Vulkan version requirements for an instance and system

Parameter Descriptions

Description

The getVulkanGraphicsRequirementsKHR function identifies to the application the minimum Vulkan version requirement and the highest known tested Vulkan version. The runtime must return ERROR_GRAPHICS_REQUIREMENTS_CALL_MISSING (ERROR_VALIDATION_FAILURE may be returned due to legacy behavior) on calls to createSession if getVulkanGraphicsRequirementsKHR has not been called for the same instance and systemId.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getVulkanGraphicsRequirementsKHR

instance must be a valid Instance handle

graphicsRequirements must be a pointer to an GraphicsRequirementsVulkanKHR structure

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

GraphicsRequirementsVulkanKHR, Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId

data GraphicsBindingVulkanKHR

XrGraphicsBindingVulkanKHR - The graphics binding structure to be passed at session creation to use Vulkan

Member Descriptions

Description

When creating a Vulkan-backed Session, the application will provide a pointer to an GraphicsBindingVulkanKHR in the next chain of the SessionCreateInfo.

Valid Usage

instance must have enabled a Vulkan API version in the range specified by GraphicsBindingVulkanKHR

instance must have enabled all the instance extensions specified by getVulkanInstanceExtensionsKHR

physicalDevice VkPhysicalDevice must match the device specified by getVulkanGraphicsDeviceKHR

device must have enabled all the device extensions specified by getVulkanDeviceExtensionsKHR

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingVulkanKHR

type must be TYPE_GRAPHICS_BINDING_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

instance must be a valid VkInstance value

physicalDevice must be a valid VkPhysicalDevice value

device must be a valid VkDevice value

See Also

StructureType, createSession

Constructors

GraphicsBindingVulkanKHR

Fields

instance' :: Ptr Instance_T

instance is a valid Vulkan VkInstance.

physicalDevice :: Ptr PhysicalDevice_T

physicalDevice is a valid Vulkan VkPhysicalDevice.

device :: Ptr Device_T

device is a valid Vulkan VkDevice.

queueFamilyIndex :: Word32

queueFamilyIndex is a valid queue family index on device.

queueIndex :: Word32

queueIndex is a valid queue index on device to be used for synchronization.

data SwapchainImageVulkanKHR

XrSwapchainImageVulkanKHR - Vulkan-specific swapchain image structure

Member Descriptions

Description

If a given session was created with GraphicsBindingVulkanKHR, the following conditions must apply.

Calls to enumerateSwapchainImages on an Swapchain in that session must return an array of SwapchainImageVulkanKHR structures.

Whenever an OpenXR function accepts an SwapchainImageBaseHeader pointer as a parameter in that session, the runtime must also accept a pointer to an SwapchainImageVulkanKHR.

The OpenXR runtime must interpret the top-left corner of the swapchain image as the coordinate origin unless specified otherwise by extension functionality.

The OpenXR runtime must interpret the swapchain images in a clip space of positive Y pointing down, near Z plane at 0, and far Z plane at 1.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SwapchainImageVulkanKHR

type must be TYPE_SWAPCHAIN_IMAGE_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

image must be a valid VkImage value

See Also

StructureType, SwapchainImageBaseHeader

Constructors

SwapchainImageVulkanKHR

Fields

image :: Image

image is a valid Vulkan VkImage to use.

data GraphicsRequirementsVulkanKHR

XrGraphicsRequirementsVulkanKHR - Vulkan API version requirements

Member Descriptions

Description

GraphicsRequirementsVulkanKHR is populated by getVulkanGraphicsRequirementsKHR with the runtime’s Vulkan API version requirements.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsRequirementsVulkanKHR

type must be TYPE_GRAPHICS_REQUIREMENTS_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrVersion, getVulkanGraphicsRequirements2KHR, getVulkanGraphicsRequirementsKHR

Constructors

GraphicsRequirementsVulkanKHR

Fields

minApiVersionSupported :: Version

minApiVersionSupported is the minimum Vulkan Instance API version that the runtime supports. Uses MAKE_VERSION on major and minor API version, ignoring any patch version component.

maxApiVersionSupported :: Version

maxApiVersionSupported is the maximum Vulkan Instance API version that the runtime has been tested on and is known to support. Newer Vulkan Instance API versions might work if they are compatible. Uses MAKE_VERSION on major and minor API version, ignoring any patch version component.

type KHR_vulkan_enable_SPEC_VERSION = 7

pattern KHR_vulkan_enable_SPEC_VERSION :: forall a. Integral a => a

type KHR_VULKAN_ENABLE_EXTENSION_NAME = "XR_KHR_vulkan_enable"

pattern KHR_VULKAN_ENABLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a