OpenXR.Extensions.XR_KHR_vulkan_enable2

Description

Name

XR_KHR_vulkan_enable2 - instance extension

Specification

See XR_KHR_vulkan_enable2 in the main specification for complete information.

Registered Extension Number

91

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

GraphicsBindingVulkan2KHR, GraphicsRequirementsVulkan2KHR, SwapchainImageVulkan2KHR, VulkanDeviceCreateInfoKHR, VulkanGraphicsDeviceGetInfoKHR, VulkanInstanceCreateInfoKHR, createVulkanDeviceKHR, createVulkanInstanceKHR, getVulkanGraphicsDevice2KHR, getVulkanGraphicsRequirements2KHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

createVulkanInstanceKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> VulkanInstanceCreateInfoKHR
createInfo extensible input struct of type XrCreateVulkanInstanceCreateInfoKHR

-> io ("vulkanInstance" ::: Ptr Instance_T, "vulkanResult" ::: Result)

xrCreateVulkanInstanceKHR - Create an OpenXR compatible VkInstance

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling createVulkanInstanceKHR

instance must be a valid Instance handle

createInfo must be a pointer to a valid VulkanInstanceCreateInfoKHR structure

vulkanInstance must be a pointer to a VkInstance value

vulkanResult must be a pointer to a VkResult value

Return Codes

Success

SUCCESS

Failure

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, VulkanInstanceCreateInfoKHR, createVulkanDeviceKHR

createVulkanDeviceKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> VulkanDeviceCreateInfoKHR
createInfo extensible input struct of type XrCreateVulkanDeviceCreateInfoKHR

-> io ("vulkanDevice" ::: Ptr Device_T, "vulkanResult" ::: Result)

xrCreateVulkanDeviceKHR - Create an OpenXR compatible VkDevice

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling createVulkanDeviceKHR

instance must be a valid Instance handle

createInfo must be a pointer to a valid VulkanDeviceCreateInfoKHR structure

vulkanDevice must be a pointer to a VkDevice value

vulkanResult must be a pointer to a VkResult value

Return Codes

Success

SUCCESS

Failure

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_HANDLE_INVALID

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, VulkanDeviceCreateInfoKHR, createVulkanInstanceKHR

getVulkanGraphicsDevice2KHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> VulkanGraphicsDeviceGetInfoKHR
getInfo extensible input struct of type VulkanGraphicsDeviceGetInfoKHR

-> io ("vulkanPhysicalDevice" ::: Ptr PhysicalDevice_T)

xrGetVulkanGraphicsDevice2KHR - Retrieve the Vulkan physical device associated with an OpenXR instance and system

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getVulkanGraphicsDevice2KHR

instance must be a valid Instance handle

getInfo must be a pointer to a valid VulkanGraphicsDeviceGetInfoKHR structure

vulkanPhysicalDevice must be a pointer to a VkPhysicalDevice value

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, VulkanGraphicsDeviceGetInfoKHR

pattern TYPE_GRAPHICS_BINDING_VULKAN2_KHR :: StructureType

pattern TYPE_SWAPCHAIN_IMAGE_VULKAN2_KHR :: StructureType

pattern TYPE_GRAPHICS_REQUIREMENTS_VULKAN2_KHR :: StructureType

getVulkanGraphicsRequirements2KHR :: MonadIO io => Instance -> SystemId -> io GraphicsRequirementsVulkanKHR

xrGetVulkanGraphicsRequirements2KHR - Retrieve the Vulkan version requirements for an instance and system

Parameters

The getVulkanGraphicsRequirementsKHR function identifies to the application the runtime’s minimum Vulkan version requirement and the highest known tested Vulkan version. getVulkanGraphicsRequirementsKHR must be called prior to calling createSession. The runtime must return ERROR_GRAPHICS_REQUIREMENTS_CALL_MISSING on calls to createSession if getVulkanGraphicsRequirements2KHR has not been called for the same instance and systemId.

Parameter Descriptions

instance is an Instance handle previously created with createInstance.

systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

graphicsRequirements is the GraphicsRequirementsVulkan2KHR output structure.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getVulkanGraphicsRequirements2KHR

instance must be a valid Instance handle

graphicsRequirements must be a pointer to an GraphicsRequirementsVulkanKHR structure

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

GraphicsRequirementsVulkan2KHR, GraphicsRequirementsVulkanKHR, Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId

data VulkanInstanceCreateInfoKHR

XrVulkanInstanceCreateInfoKHR - Vulkan Instance Create Info

Valid Usage (Implicit)

The @@ extension must be enabled prior to using VulkanInstanceCreateInfoKHR

type must be TYPE_VULKAN_INSTANCE_CREATE_INFO_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

createFlags must be 0

pfnGetInstanceProcAddr must be a valid PFN_vkGetInstanceProcAddr value

vulkanCreateInfo must be a pointer to a valid VkInstanceCreateInfo value

vulkanAllocator must be a pointer to a valid VkAllocationCallbacks value

See Also

StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, VulkanInstanceCreateFlagsKHR, createVulkanInstanceKHR

Constructors

VulkanInstanceCreateInfoKHR

Fields

systemId :: SystemId

systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

createFlags :: VulkanInstanceCreateFlagsKHR

pfnGetInstanceProcAddr :: PFN_vkGetInstanceProcAddr

pfnGetInstanceProcAddr is a function pointer to vkGetInstanceProcAddr or a compatible entry point.

vulkanCreateInfo :: Ptr (SomeStruct InstanceCreateInfo)

vulkanCreateInfo is the VkInstanceCreateInfo as specified by Vulkan.

vulkanAllocator :: Ptr AllocationCallbacks

vulkanAllocator is the VkAllocationCallbacks as specified by Vulkan.

data VulkanDeviceCreateInfoKHR

XrVulkanDeviceCreateInfoKHR - Vulkan Device Create Info

Valid Usage (Implicit)

The @@ extension must be enabled prior to using VulkanDeviceCreateInfoKHR

type must be TYPE_VULKAN_DEVICE_CREATE_INFO_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

createFlags must be 0

pfnGetInstanceProcAddr must be a valid PFN_vkGetInstanceProcAddr value

vulkanPhysicalDevice must be a valid VkPhysicalDevice value

vulkanCreateInfo must be a pointer to a valid VkDeviceCreateInfo value

vulkanAllocator must be a pointer to a valid VkAllocationCallbacks value

See Also

StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, VulkanDeviceCreateFlagsKHR, createVulkanDeviceKHR

Constructors

VulkanDeviceCreateInfoKHR

Fields

systemId :: SystemId

systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

createFlags :: VulkanDeviceCreateFlagsKHR

pfnGetInstanceProcAddr :: PFN_vkGetInstanceProcAddr

pfnGetInstanceProcAddr is a function pointer to vkGetInstanceProcAddr or a compatible entry point.

vulkanPhysicalDevice :: Ptr PhysicalDevice_T

vulkanPhysicalDevice must match getVulkanGraphicsDeviceKHR.

vulkanCreateInfo :: Ptr (SomeStruct DeviceCreateInfo)

vulkanCreateInfo is the VkDeviceCreateInfo as specified by Vulkan.

vulkanAllocator :: Ptr AllocationCallbacks

vulkanAllocator is the VkAllocationCallbacks as specified by Vulkan.

data VulkanGraphicsDeviceGetInfoKHR

XrVulkanGraphicsDeviceGetInfoKHR - Vulkan Graphics Device Get Info

Valid Usage (Implicit)

The @@ extension must be enabled prior to using VulkanGraphicsDeviceGetInfoKHR

type must be TYPE_VULKAN_GRAPHICS_DEVICE_GET_INFO_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

vulkanInstance must be a valid VkInstance value

See Also

StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId, getVulkanGraphicsDevice2KHR

Constructors

VulkanGraphicsDeviceGetInfoKHR

Fields

systemId :: SystemId

systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

vulkanInstance :: Ptr Instance_T

vulkanInstance is a valid Vulkan VkInstance.

newtype VulkanInstanceCreateFlagsKHR

XrVulkanInstanceCreateFlagsKHR - Vulkan Instance Create Info Flags

See Also

VulkanInstanceCreateInfoKHR

Constructors

VulkanInstanceCreateFlagsKHR Flags64

newtype VulkanDeviceCreateFlagsKHR

XrVulkanDeviceCreateFlagsKHR - Vulkan Device Create Info Flags

See Also

VulkanDeviceCreateInfoKHR

Constructors

VulkanDeviceCreateFlagsKHR Flags64

type GraphicsBindingVulkan2KHR = GraphicsBindingVulkanKHR

XrGraphicsBindingVulkan2KHR - The graphics binding structure to be passed at session creation to use Vulkan

Valid Usage

instance must have enabled a Vulkan API version in the range specified by getVulkanGraphicsRequirements2KHR

instance must have been created using createVulkanInstanceKHR

physicalDevice VkPhysicalDevice must match the device specified by getVulkanGraphicsDevice2KHR

device must have been created using createVulkanDeviceKHR

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingVulkan2KHR

Note: GraphicsBindingVulkan2KHR is an alias for GraphicsBindingVulkanKHR, so implicit valid usage for GraphicsBindingVulkanKHR has been replicated below.

type must be TYPE_GRAPHICS_BINDING_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

instance must be a valid VkInstance value

physicalDevice must be a valid VkPhysicalDevice value

device must be a valid VkDevice value

See Also

createSession

type SwapchainImageVulkan2KHR = SwapchainImageVulkanKHR

XrSwapchainImageVulkan2KHR - Vulkan-specific swapchain image structure

Member Descriptions

Description

If a given session was created with GraphicsBindingVulkan2KHR, the following conditions must apply.

Calls to enumerateSwapchainImages on an Swapchain in that session must return an array of SwapchainImageVulkan2KHR structures.

Whenever an OpenXR function accepts an SwapchainImageBaseHeader pointer as a parameter in that session, the runtime must also accept a pointer to an SwapchainImageVulkan2KHR.

The OpenXR runtime must interpret the top-left corner of the swapchain image as the coordinate origin unless specified otherwise by extension functionality.

The OpenXR runtime must interpret the swapchain images in a clip space of positive Y pointing down, near Z plane at 0, and far Z plane at 1.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SwapchainImageVulkan2KHR

Note: SwapchainImageVulkan2KHR is an alias for SwapchainImageVulkanKHR, so implicit valid usage for SwapchainImageVulkanKHR has been replicated below.

type must be TYPE_SWAPCHAIN_IMAGE_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

image must be a valid VkImage value

See Also

SwapchainImageBaseHeader

type GraphicsRequirementsVulkan2KHR = GraphicsRequirementsVulkanKHR

XrGraphicsRequirementsVulkan2KHR - Vulkan API version requirements

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsRequirementsVulkan2KHR

Note: GraphicsRequirementsVulkan2KHR is an alias for GraphicsRequirementsVulkanKHR, so implicit valid usage for GraphicsRequirementsVulkanKHR has been replicated below.

type must be TYPE_GRAPHICS_REQUIREMENTS_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

getVulkanGraphicsRequirements2KHR

type KHR_vulkan_enable2_SPEC_VERSION = 1

pattern KHR_vulkan_enable2_SPEC_VERSION :: forall a. Integral a => a

type KHR_VULKAN_ENABLE2_EXTENSION_NAME = "XR_KHR_vulkan_enable2"

pattern KHR_VULKAN_ENABLE2_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

data GraphicsBindingVulkanKHR

XrGraphicsBindingVulkanKHR - The graphics binding structure to be passed at session creation to use Vulkan

Member Descriptions

Description

When creating a Vulkan-backed Session, the application will provide a pointer to an GraphicsBindingVulkanKHR in the next chain of the SessionCreateInfo.

Valid Usage

instance must have enabled a Vulkan API version in the range specified by GraphicsBindingVulkanKHR

instance must have enabled all the instance extensions specified by getVulkanInstanceExtensionsKHR

physicalDevice VkPhysicalDevice must match the device specified by getVulkanGraphicsDeviceKHR

device must have enabled all the device extensions specified by getVulkanDeviceExtensionsKHR

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingVulkanKHR

type must be TYPE_GRAPHICS_BINDING_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

instance must be a valid VkInstance value

physicalDevice must be a valid VkPhysicalDevice value

device must be a valid VkDevice value

See Also

StructureType, createSession

Constructors

GraphicsBindingVulkanKHR

Fields

instance' :: Ptr Instance_T

instance is a valid Vulkan VkInstance.

physicalDevice :: Ptr PhysicalDevice_T

physicalDevice is a valid Vulkan VkPhysicalDevice.

device :: Ptr Device_T

device is a valid Vulkan VkDevice.

queueFamilyIndex :: Word32

queueFamilyIndex is a valid queue family index on device.

queueIndex :: Word32

queueIndex is a valid queue index on device to be used for synchronization.

data SwapchainImageVulkanKHR

XrSwapchainImageVulkanKHR - Vulkan-specific swapchain image structure

Member Descriptions

Description

If a given session was created with GraphicsBindingVulkanKHR, the following conditions must apply.

Calls to enumerateSwapchainImages on an Swapchain in that session must return an array of SwapchainImageVulkanKHR structures.

Whenever an OpenXR function accepts an SwapchainImageBaseHeader pointer as a parameter in that session, the runtime must also accept a pointer to an SwapchainImageVulkanKHR.

The OpenXR runtime must interpret the top-left corner of the swapchain image as the coordinate origin unless specified otherwise by extension functionality.

The OpenXR runtime must interpret the swapchain images in a clip space of positive Y pointing down, near Z plane at 0, and far Z plane at 1.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using SwapchainImageVulkanKHR

type must be TYPE_SWAPCHAIN_IMAGE_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

image must be a valid VkImage value

See Also

StructureType, SwapchainImageBaseHeader

Constructors

SwapchainImageVulkanKHR

Fields

image :: Image

image is a valid Vulkan VkImage to use.

data GraphicsRequirementsVulkanKHR

XrGraphicsRequirementsVulkanKHR - Vulkan API version requirements

Member Descriptions

Description

GraphicsRequirementsVulkanKHR is populated by getVulkanGraphicsRequirementsKHR with the runtime’s Vulkan API version requirements.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsRequirementsVulkanKHR

type must be TYPE_GRAPHICS_REQUIREMENTS_VULKAN_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

See Also

StructureType, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrVersion, getVulkanGraphicsRequirements2KHR, getVulkanGraphicsRequirementsKHR

Constructors

GraphicsRequirementsVulkanKHR

Fields

minApiVersionSupported :: Version

minApiVersionSupported is the minimum Vulkan Instance API version that the runtime supports. Uses MAKE_VERSION on major and minor API version, ignoring any patch version component.

maxApiVersionSupported :: Version

maxApiVersionSupported is the maximum Vulkan Instance API version that the runtime has been tested on and is known to support. Newer Vulkan Instance API versions might work if they are compatible. Uses MAKE_VERSION on major and minor API version, ignoring any patch version component.

getVulkanGraphicsRequirementsKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> SystemId
systemId is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId handle for the system which will be used to create a session.

-> io GraphicsRequirementsVulkanKHR

xrGetVulkanGraphicsRequirementsKHR - Retrieve the Vulkan version requirements for an instance and system

Parameter Descriptions

Description

The getVulkanGraphicsRequirementsKHR function identifies to the application the minimum Vulkan version requirement and the highest known tested Vulkan version. The runtime must return ERROR_GRAPHICS_REQUIREMENTS_CALL_MISSING (ERROR_VALIDATION_FAILURE may be returned due to legacy behavior) on calls to createSession if getVulkanGraphicsRequirementsKHR has not been called for the same instance and systemId.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling getVulkanGraphicsRequirementsKHR

instance must be a valid Instance handle

graphicsRequirements must be a pointer to an GraphicsRequirementsVulkanKHR structure

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_SYSTEM_INVALID

ERROR_VALIDATION_FAILURE

ERROR_FUNCTION_UNSUPPORTED

See Also

GraphicsRequirementsVulkanKHR, Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrSystemId