OpenXR.Extensions.XR_KHR_vulkan_swapchain_format_list

Description

Name

XR_KHR_vulkan_swapchain_format_list - instance extension

Specification

See XR_KHR_vulkan_swapchain_format_list in the main specification for complete information.

Registered Extension Number

15

Revision

3

Extension and Version Dependencies

Requires OpenXR 1.0

Requires @@

See Also

VulkanSwapchainFormatListCreateInfoKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data VulkanSwapchainFormatListCreateInfoKHR

XrVulkanSwapchainFormatListCreateInfoKHR - A list of Vulkan view formats

Valid Usage (Implicit)

The @@ extension must be enabled prior to using VulkanSwapchainFormatListCreateInfoKHR

type must be TYPE_VULKAN_SWAPCHAIN_FORMAT_LIST_CREATE_INFO_KHR

next must be NULL or a valid pointer to the next structure in a structure chain

If viewFormatCount is not 0, viewFormats must be a pointer to an array of viewFormatCount valid VkFormat values

See Also

StructureType, createSwapchain

Constructors

VulkanSwapchainFormatListCreateInfoKHR

Fields

viewFormats :: Vector Format

viewFormats is an array of VkFormat.

type KHR_vulkan_swapchain_format_list_SPEC_VERSION = 3

pattern KHR_vulkan_swapchain_format_list_SPEC_VERSION :: forall a. Integral a => a

type KHR_VULKAN_SWAPCHAIN_FORMAT_LIST_EXTENSION_NAME = "XR_KHR_vulkan_swapchain_format_list"

pattern KHR_VULKAN_SWAPCHAIN_FORMAT_LIST_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a