OpenXR.Extensions.XR_KHR_win32_convert_performance_counter_time

Description

Name

XR_KHR_win32_convert_performance_counter_time - instance extension

Specification

See XR_KHR_win32_convert_performance_counter_time in the main specification for complete information.

Registered Extension Number

36

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

convertTimeToWin32PerformanceCounterKHR, convertWin32PerformanceCounterToTimeKHR

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

convertTimeToWin32PerformanceCounterKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> Time
time is an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime.

-> io ("performanceCounter" ::: LARGE_INTEGER)

xrConvertTimeToWin32PerformanceCounterKHR - Convert XrTime to Win32 QueryPerformanceCounter time

Parameter Descriptions

Description

The convertTimeToWin32PerformanceCounterKHR function converts an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime to time as if generated by the QueryPerformanceCounter Windows function.

If the output performanceCounter cannot represent the input time, the runtime must return ERROR_TIME_INVALID.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling convertTimeToWin32PerformanceCounterKHR

instance must be a valid Instance handle

performanceCounter must be a pointer to a LARGE_INTEGER value

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_VALIDATION_FAILURE

ERROR_TIME_INVALID

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime

convertWin32PerformanceCounterToTimeKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is an Instance handle previously created with createInstance.

-> ("performanceCounter" ::: LARGE_INTEGER)
performanceCounter is a time returned by QueryPerformanceCounter.

-> io Time

xrConvertWin32PerformanceCounterToTimeKHR - Convert Win32 QueryPerformanceCounter time to XrTime

Parameter Descriptions

Description

The convertWin32PerformanceCounterToTimeKHR function converts a time stamp obtained by the QueryPerformanceCounter Windows function to the equivalent https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime.

If the output time cannot represent the input performanceCounter, the runtime must return ERROR_TIME_INVALID.

Valid Usage (Implicit)

The @@ extension must be enabled prior to calling convertWin32PerformanceCounterToTimeKHR

instance must be a valid Instance handle

performanceCounter must be a pointer to a valid LARGE_INTEGER value

time must be a pointer to an https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime value

Return Codes

Success

SUCCESS

Failure

ERROR_HANDLE_INVALID

ERROR_INSTANCE_LOST

ERROR_RUNTIME_FAILURE

ERROR_VALIDATION_FAILURE

ERROR_TIME_INVALID

ERROR_FUNCTION_UNSUPPORTED

See Also

Instance, https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XrTime

type KHR_win32_convert_performance_counter_time_SPEC_VERSION = 1

pattern KHR_win32_convert_performance_counter_time_SPEC_VERSION :: forall a. Integral a => a

type KHR_WIN32_CONVERT_PERFORMANCE_COUNTER_TIME_EXTENSION_NAME = "XR_KHR_win32_convert_performance_counter_time"

pattern KHR_WIN32_CONVERT_PERFORMANCE_COUNTER_TIME_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type LARGE_INTEGER = Word64