OpenXR.Extensions.XR_MNDX_egl_enable

Description

Name

XR_MNDX_egl_enable - instance extension

Specification

See XR_MNDX_egl_enable in the main specification for complete information.

Registered Extension Number

49

Revision

1

Extension and Version Dependencies

Requires OpenXR 1.0

See Also

GraphicsBindingEGLMNDX

Document Notes

For more information, see the OpenXR Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data GraphicsBindingEGLMNDX

XrGraphicsBindingEGLMNDX - The graphics binding structure to be passed at session creation to EGL

Member Descriptions

Description

When creating an EGL based Session, the application will provide a pointer to an GraphicsBindingEGLMNDX structure in the next chain of the SessionCreateInfo.

The required window system configuration define to expose this structure type is XR_USE_PLATFORM_EGL.

Valid Usage (Implicit)

The @@ extension must be enabled prior to using GraphicsBindingEGLMNDX

type must be TYPE_GRAPHICS_BINDING_EGL_MNDX

next must be NULL or a valid pointer to the next structure in a structure chain

getProcAddress must be a valid PFNEGLGETPROCADDRESSPROC value

display must be a valid EGLDisplay value

config must be a valid EGLConfig value

context must be a valid EGLContext value

See Also

StructureType, createSession

Constructors

GraphicsBindingEGLMNDX

Fields

getProcAddress :: PFNEGLGETPROCADDRESSPROC

getProcAddress is a valid function pointer to eglGetProcAddress.

display :: EGLDisplay

display is a valid EGL EGLDisplay.

config :: EGLConfig

config is a valid EGL EGLConfig.

context :: EGLContext

context is a valid EGL EGLContext.

type MNDX_egl_enable_SPEC_VERSION = 1

pattern MNDX_egl_enable_SPEC_VERSION :: forall a. Integral a => a

type MNDX_EGL_ENABLE_EXTENSION_NAME = "XR_MNDX_egl_enable"

pattern MNDX_EGL_ENABLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type PFNEGLGETPROCADDRESSPROC = FunPtr (CString -> IO (FunPtr (IO ())))

type EGLDisplay = Ptr ()

type EGLConfig = Ptr ()

type EGLContext = Ptr ()